% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{add_bags_if_not_exist}
\alias{add_bags_if_not_exist}
\title{Add new bags to the database if they don't exist}
\usage{
add_bags_if_not_exist(df.payload, con, manifest_name_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{manifest_name_col}{Column name in df.payload for manifest name.}
}
\value{
A result from the dbAppendTable indicating if the bags were added successfully.
}
\description{
Add new bags to the database if they don't exist
}
