% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{add_malaria_blood_controls_if_not_exist}
\alias{add_malaria_blood_controls_if_not_exist}
\title{Add malaria blood controls to the database if they don't exist}
\usage{
add_malaria_blood_controls_if_not_exist(df.payload, con, density_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{density}{User Data Density Column}
}
\value{
A result from the dbAppendTable indicating if the controls were added successfully.
}
\description{
Add malaria blood controls to the database if they don't exist
}
