% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{add_new_dbs_control_sheets}
\alias{add_new_dbs_control_sheets}
\title{Add new DBS control sheets to the database}
\usage{
add_new_dbs_control_sheets(con, df.payload, dbs_sheet_name_col)
}
\arguments{
\item{con}{A database connection object.}

\item{df.payload}{A dataframe with payload data.}

\item{dbs_sheet_name_col}{Column name in df.payload for DBS Sheet name.}
}
\value{
A result from the dbAppendTable indicating if the sheets were added successfully.
}
\description{
Add new DBS control sheets to the database
}
