% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{add_position_column}
\alias{add_position_column}
\title{Add a position column based on column attributes}
\usage{
add_position_column(df, col_attributes)
}
\arguments{
\item{df}{A dataframe to modify.}

\item{col_attributes}{A named list of column attributes for exceptions.}
}
\value{
A dataframe, potentially with a new 'position' column.
}
\description{
Add a position column based on column attributes
}
\keyword{utility}
