% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{add_row_numbers}
\alias{add_row_numbers}
\title{Add row numbers to a dataframe}
\usage{
add_row_numbers(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe with an additional 'RowNumber' column.
}
\description{
Add row numbers to a dataframe
}
\keyword{utility}
