% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{add_to_errors}
\alias{add_to_errors}
\title{Add an Error to the Error List}
\usage{
add_to_errors(errors, err)
}
\arguments{
\item{errors}{A list object that contains the current errors.}

\item{err}{An error object to be added to the error list. It can be NULL, in which case no action is taken.}
}
\value{
The updated errors list.
}
\description{
This utility function checks if an error exists and, if so, appends it to the provided error list.
}
\examples{
\dontrun{
  errors <- list()  
  new_error <- list(description = "Some error")  # Hypothetical error structure.
  errors <- add_to_errors(errors, new_error)
}
}
