% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{append_boxes_if_not_exist}
\alias{append_boxes_if_not_exist}
\title{Process Whole Blood Box Data}
\usage{
append_boxes_if_not_exist(
  con,
  user_data,
  created_col,
  last_updated_col,
  box_name_col,
  box_barcode_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A dataframe containing the payload data.}

\item{created_col}{Column name in user_data for created date.}

\item{last_updated_col}{Column name in user_data for last updated date.}

\item{box_name_col}{Column name in user_data for box name.}

\item{box_barcode_col}{Column name in user_data for box barcode.}
}
\value{
A result from the dbAppendTable indicating if the boxes were added successfully.
}
\description{
This function will check for locations and add new boxes if they don't exist. It will return the boxes and locations IDs.
}
\keyword{internal}
