% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateReferences.R
\name{append_strains_to_db}
\alias{append_strains_to_db}
\title{Append Strains to the Database Table}
\usage{
append_strains_to_db(user_file, database = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{user_file}{A data frame containing a column named 'strain'.}

\item{database}{The path to the SQLite database.}
}
\value{
A boolean indicating success (TRUE) or failure (FALSE).
}
\description{
This function appends new strains to the 'strain' table in the database.
}
\examples{
# Assuming you have a data frame named user_data with a strain column and a valid SQLite database at "/path/to/db"
result <- append_strains_to_db(user_data, "/path/to/db")

}
