% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{append_whole_blood_tubes}
\alias{append_whole_blood_tubes}
\title{Add whole blood tubes to the database if they don't exist}
\usage{
append_whole_blood_tubes(con, user_data, barcode_col, position_col)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A dataframe containing the payload data.}

\item{barcode_col}{Column name in user_data for barcode.}

\item{position_col}{Column name in user_data for position.}
}
\value{
A result from the dbAppendTable indicating if the tubes were added successfully.
}
\description{
Add whole blood tubes to the database if they don't exist
}
