% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAndDeleteSamples.R
\name{ArchiveAndDeleteSamples}
\alias{ArchiveAndDeleteSamples}
\title{Archive and Delete Samples}
\usage{
ArchiveAndDeleteSamples(
  operation,
  data,
  comment,
  status_id,
  verification = TRUE
)
}
\value{
None
}
\description{
Archive and Delete Samples
}
