% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_SampleDB.R
\name{Backup_SampleDB}
\alias{Backup_SampleDB}
\title{Backup Sample Database}
\usage{
Backup_SampleDB(
  database = Sys.getenv("SDB_PATH"),
  backup_dir = NULL,
  checksum = TRUE
)
}
\arguments{
\item{database}{Path to the sample database, defaults to the "SDB_PATH" environment variable.}

\item{backup_dir}{Directory to store the backups; if NULL, it defaults to a subdirectory 'backups' in the database directory.}

\item{checksum}{Logical; if TRUE, checks if the new backup is different from the last one.}
}
\description{
Backs up the sample database and optionally compares it to the latest backup to detect changes.
The function automatically compresses the backup file.
}
