% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{bind_new_data}
\alias{bind_new_data}
\title{Add Data to User Data}
\usage{
bind_new_data(df, named_list)
}
\arguments{
\item{df}{The dataframe to which the data will be added.}

\item{named_list}{The named list whose elements should be added as a new column to the dataframe.}
}
\value{
The dataframe with the added column.
}
\description{
This function takes a named list and appends its values as a new column to a dataframe.
The name of the named list is matched with the dataframe's column names.
}
\examples{
df <- data.frame(Name = c("John", "Doe"), Age = c(25, 30))
new_data <- list(Designation = "Manager")
df <- bind_new_data(df, new_data)

}
