% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{check_control_exists}
\alias{check_control_exists}
\title{Check for Control Existence}
\usage{
check_control_exists(
  con,
  table_name,
  row_number_col,
  control_col,
  batch_col,
  error_if_exists = FALSE
)
}
\arguments{
\item{con}{A database connection.}

\item{table_name}{The name of the table to check.}

\item{row_number_col}{The name of the row number column.}

\item{control_col}{The name of the control column.}

\item{batch_col}{The name of the batch column.}

\item{error_if_exists}{If TRUE, returns an error if the control ID exists. If FALSE,
returns an error if the control ID does not exist.}
}
\value{
NULL if there is no error, otherwise an ErrorData object with
appropriate description and data frame.
}
\description{
This function checks for unmatched rows based on the control IDs
in a specified table when compared to the \code{study_subject} table.
It returns an error data if there are unmatched rows.
}
