% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_cryovial_barcodes_exist}
\alias{check_cryovial_barcodes_exist}
\title{Check Cryovial Barcodes in Database}
\usage{
check_cryovial_barcodes_exist(
  con,
  user_data,
  row_number_col,
  cryovial_col,
  error_if_exists = TRUE
)

check_cryovial_barcodes_exist(
  con,
  user_data,
  row_number_col,
  cryovial_col,
  error_if_exists = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the table where the user data is temporarily stored in the database.}

\item{row_number_col}{The column with the row number in the \code{user_data}.}

\item{cryovial_col}{The column with Cryovial barcodes in the \code{user_data}.}

\item{error_if_exists}{Logical. If TRUE, an error is returned if the barcode exists in the database.}
}
\value{
An instance of the ErrorData class if errors are found, or NULL if there are no errors.

An instance of the ErrorData class or NULL.
}
\description{
This function checks whether Cryovial barcodes provided in a user-uploaded table
exist or don't exist (based on \code{error_if_exists} parameter) in the main database.
}
\keyword{cryovial}
\keyword{validation,}
