% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_cryovial_box_exists}
\alias{check_cryovial_box_exists}
\title{Check If Cryovial Box Exists in the Database}
\usage{
check_cryovial_box_exists(
  con,
  table_name,
  row_number_col,
  box_name_col,
  box_barcode_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the table containing user-uploaded data.}

\item{row_number_col}{The name of the column in the user-uploaded table that contains row numbers.}

\item{box_name_col}{The name of the column in the user-uploaded table that contains the cryovial container names.}

\item{box_barcode_col}{The name of the column in the user-uploaded table that contains the manifest barcodes linked with the cryovial.}
}
\value{
An instance of the ErrorData class if errors are found, or NULL if there are no errors.
}
\description{
This function checks whether the cryovial containers provided in a user-uploaded table
exist in the main cryovial tube database.
}
\keyword{cryovial}
\keyword{validation,}
