% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{check_duplicated_rows}
\alias{check_duplicated_rows}
\title{Check for duplicated rows in a data frame}
\usage{
check_duplicated_rows(formatted_csv, col_attributes = NULL)
}
\arguments{
\item{formatted_csv}{The user-provided data frame to check for duplicated rows.}

\item{col_attributes}{A named list of expected column attributes. If you want to check only specific columns for duplicates, include their names here. Otherwise, it will check all columns.}
}
\value{
An ErrorData object if any duplicated rows are found, otherwise NULL.
}
\description{
Check for duplicated rows in a data frame
}
