% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{check_for_duplicates_in_a_row}
\alias{check_for_duplicates_in_a_row}
\title{Check for duplicates in row}
\usage{
check_for_duplicates_in_a_row(user_data, row_number)
}
\arguments{
\item{user_data}{The user data data.frame.}

\item{row_number}{Numerical value indicating which row to check for duplicates}
}
\value{
A list with the column names that were found more than once.
}
\description{
This function will check if there are duplicated values in a row. This function
is used to identify a if a users file has duplicated column headers before
setting the column names.
}
\keyword{internal}
