% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_longitudinal_study_dates}
\alias{check_longitudinal_study_dates}
\title{Check if collection dates are missing for samples in a longitudinal study}
\usage{
check_longitudinal_study_dates(
  con,
  table_name,
  row_number_col,
  study_short_code_col,
  collection_date_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the table where the user data is temporarily stored in the database.}

\item{row_number_col}{The name of the column containing the row number.}

\item{study_short_code_col}{The name of the column containing the study short code.}

\item{collection_date_col}{The name of the column containing the collection date.}
}
\value{
An instance of the ErrorData class or NULL.
}
\description{
Check if collection dates are missing for samples in a longitudinal study
}
\keyword{validation}
