% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_micronix_barcodes_exist}
\alias{check_micronix_barcodes_exist}
\title{Check Micronix Barcodes in Database}
\usage{
check_micronix_barcodes_exist(
  con,
  user_data,
  row_number_col,
  micronix_col,
  error_if_exists = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the table containing user-uploaded data.}

\item{row_number_col}{The name of the column in the user-uploaded table that contains row numbers.}

\item{micronix_col}{The name of the column in the user-uploaded table that contains the Micronix barcodes.}

\item{error_if_exists}{A logical value. If TRUE, function returns error if barcode exists in the database,
if FALSE, function returns error if barcode doesn't exist.}
}
\value{
An instance of the ErrorData class if errors are found, or NULL if there are no errors.
}
\description{
This function checks whether Micronix barcodes provided in a user-uploaded table
exist or don't exist (based on \code{error_if_exists} parameter) in the main database.
}
\keyword{micronix}
\keyword{validation,}
