% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_micronix_plate_exists}
\alias{check_micronix_plate_exists}
\title{Check If Micronix Plate Exists in the Database}
\usage{
check_micronix_plate_exists(
  con,
  table_name,
  row_number_col,
  plate_name_col,
  plate_barcode_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the table containing user-uploaded data.}

\item{row_number_col}{The name of the column in the user-uploaded table that contains row numbers.}

\item{plate_name_col}{The name of the column in the user-uploaded table that contains the Micronix container names.}

\item{plate_barcode_col}{The name of the column in the user-uploaded table that contains the Micronix container barcodes.}
}
\value{
An instance of the ErrorData class if errors are found, or NULL if there are no errors.
}
\description{
This function checks whether the Micronix plates provided in a user-uploaded table
exist in the main micronix plate database.
}
\keyword{micronix}
\keyword{validation,}
