% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{check_missing_data}
\alias{check_missing_data}
\title{Check for missing data in required fields}
\usage{
check_missing_data(formatted_csv, col_attributes)
}
\arguments{
\item{formatted_csv}{The user provided dataframe to check.}

\item{col_attributes}{A named list of expected column attributes.}
}
\value{
An ErrorData object if any required data is missing, otherwise NULL.
}
\description{
Check for missing data in required fields
}
