% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{check_strain_percentage_match}
\alias{check_strain_percentage_match}
\title{Validate that the strain and percentage columns have matching semicolon counts}
\usage{
check_strain_percentage_match(data, percentage_col, strain_col)
}
\arguments{
\item{data}{The users data.}

\item{percentage_col}{The name of the column containing percentages.}

\item{strain_col}{The name of the column containing strains.}
}
\value{
NULL or ErrorData.
}
\description{
Validate that the strain and percentage columns have matching semicolon counts
}
\keyword{compositions}
\keyword{pre-validation}
\keyword{strains}
