% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{check_unique_positions}
\alias{check_unique_positions}
\title{Check Unique Positions}
\usage{
check_unique_positions(
  user_data,
  position_col,
  container_name_col,
  container_barcode_col
)
}
\arguments{
\item{user_data}{A data frame contianer the user data}

\item{position_col}{A position column}

\item{container_name_col}{The container name column}

\item{container_barcode_col}{The container barcode column}
}
\description{
Check Unique Positions
}
