% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{collapse_and_nest_columns}
\alias{collapse_and_nest_columns}
\title{Collapse and Nest Strain and Percentage Columns}
\usage{
collapse_and_nest_columns(
  user_data,
  row_number_col,
  strains_col,
  percentage_col
)
}
\arguments{
\item{user_data}{A data frame with the specified strain and percentage columns.}

\item{strains_col}{The column name in the data frame containing strain data.}

\item{percentage_col}{The column name in the data frame containing percentage data.}
}
\value{
A data frame with nested strain and percentage columns.
}
\description{
This function takes in a data frame and collapses the provided strain and percentage columns using a semicolon (;).
The individual strains and percentages in each row are combined into single, semicolon-separated strings.
This is commonly used after manipulating or analyzing strains and their corresponding compositions.
}
\examples{
\dontrun{
result <- collapse_and_nest_columns(df, "strain_column_name", "percentage_column_name")
}

}
\keyword{compositions}
\keyword{post-processing}
\keyword{strains}
