% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{ColumnData}
\alias{ColumnData}
\title{Create a ColumnData S3 object}
\usage{
ColumnData(required = NULL, conditional = NULL, optional = NULL)
}
\arguments{
\item{required}{A vector or NULL. Represents required columns.}

\item{conditional}{A vector or NULL. Represents conditional columns.}

\item{optional}{A vector or NULL. Represents optional columns.}
}
\value{
An S3 object of class 'ColumnData'.
}
\description{
This function serves as a constructor for the ColumnData S3 class.
It is used to represent column information in a structured manner.
}
