% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{concat_position}
\alias{concat_position}
\title{Concatenate Position from Column and Row Information}
\usage{
concat_position(col, row)
}
\arguments{
\item{col}{A character vector indicating the column.}

\item{row}{A numeric vector indicating the row.}
}
\value{
A character vector of the concatenated position.
}
\description{
This function takes a column (character) and a row (numeric) and returns a formatted string by concatenating the column and row values.
}
\examples{
concat_position("A", 5) # "A05"
concat_position("B", 1) # "B01"
concat_position("AA", 23) # "AA23"
}
