% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{convert_density_to_numeric}
\alias{convert_density_to_numeric}
\title{Convert string values with 'k' or 'K' to numeric values}
\usage{
convert_density_to_numeric(data)
}
\arguments{
\item{data}{A data frame with a column named 'density'.}
}
\value{
A list with the processed data and potential ErrorData.
}
\description{
Convert string values with 'k' or 'K' to numeric values
}
\keyword{compositions}
