% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{copy_database}
\alias{copy_database}
\title{Copy SQLite Database}
\usage{
copy_database(database, new_database)
}
\arguments{
\item{database}{Path to the original SQLite database.}

\item{new_database}{Path to the destination for the SQLite database copy.}
}
\value{
TRUE if the copy is successful, otherwise FALSE.
}
\description{
This function copies an existing SQLite database to a new location.
}
\examples{
\dontrun{
  success <- copy_database("path_to_original.db", "path_to_new.db")
  if (!success) {
    stop("Failed to copy the database.")
  }
}
}
