% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{copy_if_not_exists}
\alias{copy_if_not_exists}
\title{Copy a file to a destination if it does not exist there.}
\usage{
copy_if_not_exists(src, dest)
}
\arguments{
\item{src}{The source file path.}

\item{dest}{The destination file path.}
}
\value{
Returns TRUE if the file is copied, FALSE otherwise.
}
\description{
Copy a file to a destination if it does not exist there.
}
\examples{
copy_if_not_exists("source.txt", "destination.txt")

}
