% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{create_control_label}
\alias{create_control_label}
\title{Create Control Label with Density and Composition Values}
\usage{
create_control_label(dens_values, comp_values)
}
\arguments{
\item{dens_values}{A numeric vector specifying the density values.}

\item{comp_values}{A character vector specifying the composition values.}
}
\value{
A character vector with the control labels.
}
\description{
This function takes vectors for density and composition and creates a control label
by combining the values. Densities greater than 999 are divided
by 1000 and appended with a 'K'.
}
\examples{
dens <- c(100, 1000, 1500)
comp <- c("S1_1", "S3_2", "S2_2")
create_control_label(dens, comp)

}
