% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{create_controls_for_batch}
\alias{create_controls_for_batch}
\title{Create Controls for the batch}
\usage{
create_controls_for_batch(
  user_data,
  con,
  density_col,
  study_short_code_col,
  control_col
)
}
\arguments{
\item{user_data}{A dataframe containing user data to process.}

\item{con}{A database connection object.}

\item{density_col}{Column name in user_data for density value.}

\item{study_short_code_col}{Name of the column in user_data corresponding to "Batch".}

\item{control_col}{Name of the column in user-dta correspond to "Control" (system made)}
}
\value{
A dataframe with the study_subject_id added.
}
\description{
This function prepares user data, appends study subjects to a database table, and rejoins to get the study_subject_id.
}
