% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{create_flat_view_by_sample_type}
\alias{create_flat_view_by_sample_type}
\title{Flatten all database tables by sample type}
\usage{
create_flat_view_by_sample_type(
  sample_storage_type,
  database = Sys.getenv("SDB_PATH")
)
}
\arguments{
\item{sample_storage_type}{Samples to view by type. Options are 'micronix' and 'cryovial'.}

\item{database}{The path to the database.}
}
\description{
This function is a utility function to pull all samples from the database.
}
