% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{create_flat_view_of_extractions}
\alias{create_flat_view_of_extractions}
\title{View all extractions in the database}
\usage{
create_flat_view_of_extractions(database = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{database}{The path to the database.}
}
\description{
This function is a utility function to pull all extractions from the database.
}
