% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{create_unique_id}
\alias{create_unique_id}
\title{Create a unique identifier for compositions}
\usage{
create_unique_id(strains, percentages)
}
\arguments{
\item{strains}{A character vector containing strain names.}

\item{percentages}{A numeric vector containing percentages for each strain.}
}
\value{
A character string representing the unique identifier for the composition.
}
\description{
This function creates a unique identifier for given strains and their respective percentages.
The strains and percentages are sorted based on percentages to ensure consistency in unique ID creation.
}
\examples{
strains <- c("D6", "3D7", "HB3")
percentages <- c(0.32, 0.33, 0.35)
create_unique_id(strains, percentages)
}
