% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{create_unique_id_from_sorted}
\alias{create_unique_id_from_sorted}
\title{Create a unique identifier from sorted strains and percentages}
\usage{
create_unique_id_from_sorted(sorted_strains, sorted_percentages)
}
\arguments{
\item{sorted_strains}{A character vector of sorted strains.}

\item{sorted_percentages}{A numeric vector of sorted percentages.}
}
\value{
A character string representing the unique identifier.
}
\description{
This function generates a unique identifier by combining sorted strains and percentages.
}
