% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cryovial_na}
\alias{cryovial_na}
\title{Cryovial NA}
\format{
\subsection{\code{micronix_na}}{

A data frame with 1 rows and 13 columns:
\describe{
\item{Barcode}{A 10 digit unique identifier for the sample}
\item{BoxRow}{A character denoting the row position of the sample}
\item{BoxColumn}{An integer denoting the column position of the sample}
\item{SpecimenType}{The specimen contained in the sample}
\item{StudySubject}{The identifier of the study subject from which the sample was collected}
\item{CollectionDate}{The date the sample was collected. This field is required if the sample is part of a \verb{longitudinal study}, and MUST be in \code{YYYY-MM-DD} format}
\item{BoxName}{The name of the plate}
\item{FreezerName}{The name of the freezer the sample will be in}
\item{RackNumber}{The shelf the sample will be located in}
\item{RackPosition}{The place on the shelf where the sample will be located}
\item{StudyCode}{The code associated with the study the sample is part of}
\item{Comment}{An optional field to leave a comment for the sample}
\item{BoxBarcode}{An optional field to provide the barcode of the container the sample is located in}
...
}
}
}
\source{
EPPIcenter
}
\usage{
cryovial_na
}
\description{
An example upload file with a single Cryovial sample using the NA file format
}
\keyword{datasets}
