% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteEmptyContainer.R
\name{DeleteEmptyContainer}
\alias{DeleteEmptyContainer}
\title{Delete Empty Wetlab Containers in the EPPIcenter SampleDB database}
\usage{
DeleteEmptyContainer(database, type, container_name)
}
\description{
Delete Empty Wetlab Containers in the EPPIcenter SampleDB database
}
