% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteEmptyContainer.R
\name{DeleteEmptyContainers}
\alias{DeleteEmptyContainers}
\title{Delete Empty Wetlab Containers in the EPPIcenter SampleDB database}
\usage{
DeleteEmptyContainers(database, type, container_name)
}
\description{
Delete Empty Wetlab Containers in the EPPIcenter SampleDB database
}
