% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAndDeleteSamples.R
\name{DeleteWholeBloodSamples}
\alias{DeleteWholeBloodSamples}
\title{Delete Whole Blood Controls}
\usage{
DeleteWholeBloodSamples(whole_blood_tube_ids)
}
\arguments{
\item{whole_blood_tube_ids}{A vector containing the IDs of whole blood tubes to be deleted.}
}
\value{
None
}
\description{
This function deletes whole blood controls from the database based on the provided list of whole_blood_tube IDs.
It also handles the deletion of associated records if a whole blood sample is the last one for a study subject.
}
\examples{
\dontrun{
# Delete whole blood controls with IDs 1, 2, and 3
DeleteWholeBloodSamples(c(1, 2, 3))
}

}
