% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{denormalize_composition_ids}
\alias{denormalize_composition_ids}
\title{Denormalize Composition IDs from User Data}
\usage{
denormalize_composition_ids(
  user_data,
  in_label_col,
  in_index_col,
  in_legacy_col,
  out_composition_id_col
)
}
\arguments{
\item{user_data}{A dataframe containing user data with split composition IDs.}

\item{in_label_col}{A string specifying the column name of the labels in \code{user_data}. Default is "Label".}

\item{in_index_col}{A string specifying the column name of the indexes in \code{user_data}. Default is "Index".}

\item{in_legacy_col}{A string specifying the column name indicating if the composition ID is legacy in \code{user_data}. Default is "LegacyLabel".}

\item{out_composition_id_col}{A string specifying the name of the new column where the combined composition IDs will be stored.}
}
\value{
A dataframe with a new column (specified by composition_id_col) containing the combined composition ID.
}
\description{
This function processes a dataframe to combine the separate columns (Label, Index, and LegacyLabel)
back into a single composition ID column based on their format.
}
