% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{deploy_or_notify_existing_app}
\alias{deploy_or_notify_existing_app}
\title{Deploy the Shiny app or notify if already deployed.}
\usage{
deploy_or_notify_existing_app(shiny_server_path, pkgname, pkg_path)
}
\arguments{
\item{shiny_server_path}{Path to the Shiny server.}

\item{pkgname}{Name of the package.}

\item{pkg_path}{Path of the package installation.}
}
\description{
Deploy the Shiny app or notify if already deployed.
}
\keyword{internal}
