% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{deploy_shiny_app}
\alias{deploy_shiny_app}
\title{Deploy Shiny Application}
\usage{
deploy_shiny_app(pkgname, site_install)
}
\arguments{
\item{pkgname}{The name of the package containing the Shiny application to be deployed.}

\item{site_install}{Logical indicating if this is a site installation.}
}
\description{
This function checks if the environment meets the requirements to deploy
a Shiny application and attempts to deploy it if all conditions are met.
}
\examples{
\dontrun{
deploy_shiny_app("my_package", TRUE)
}

}
\keyword{internal}
