% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{detect_missing_columns}
\alias{detect_missing_columns}
\title{Detect Missing Columns in User's CSV File}
\usage{
detect_missing_columns(
  user_file,
  column_attributes,
  database = Sys.getenv("SDB_PATH")
)
}
\arguments{
\item{user_file}{A data frame representing the user's CSV file.}

\item{column_attributes}{An object containing attributes for columns like required, conditional, and action.}

\item{database}{A character string specifying the path to the SQLite database.}
}
\value{
A character vector of missing columns from the user's CSV file.
}
\description{
This function checks for any required columns that are missing from the user's CSV file.
Additionally, it checks special conditions such as the presence of "CollectionDate" for
specific studies that are longitudinal.
}
\examples{
\dontrun{
user_data <- data.frame(StudyCode = c("ST001"), StudySubject = c("SS001"))
column_attrs <- FileColumnAttributes$new()
database_path <- "path_to_database.sqlite"
detect_missing_columns(user_data, column_attrs, database_path)
}
}
\keyword{internal}
