% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_SampleDB.R
\name{.Backup}
\alias{.Backup}
\title{Internal function to backup a SQLite database}
\usage{
.Backup(database, backup_dest)
}
\arguments{
\item{database}{Path to the SQLite database file.}

\item{backup_dest}{Destination path for the backup file.}
}
\value{
Returns TRUE if the backup file exists, FALSE otherwise.
}
\description{
Internal function to backup a SQLite database
}
