% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_SampleDB.R
\name{.Compress}
\alias{.Compress}
\title{Internal function to compress a file}
\usage{
.Compress(source_file, destination_file)
}
\arguments{
\item{source_file}{The original file to be compressed.}

\item{destination_file}{The destination path for the compressed file.}
}
\value{
The path to the compressed file.
}
\description{
Internal function to compress a file
}
