% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_SampleDB.R
\name{.Uncompress}
\alias{.Uncompress}
\title{Internal function to uncompress a file}
\usage{
.Uncompress(source_file, destination_file)
}
\arguments{
\item{source_file}{Compressed file path.}

\item{destination_file}{Destination path for the uncompressed file.}
}
\value{
Returns TRUE if the destination file exists after decompression.
}
\description{
Internal function to uncompress a file
}
