% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSpecimens.R
\name{.UpdateActiveControlCount}
\alias{.UpdateActiveControlCount}
\title{Update the Number of Extracted DNA Controls After Extraction}
\usage{
.UpdateActiveControlCount(user_data, control_extraction, con)
}
\arguments{
\item{user_data}{A data frame containing the DNA extraction data.}

\item{control_extraction}{A character string specifying the type of control extraction.}

\item{con}{A database connection object.}
}
\description{
This internal function updates the number of active controls for a given type
of DNA extraction after the extraction has been performed.
}
\keyword{internal}
