% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{ErrorDataList}
\alias{ErrorDataList}
\title{ErrorDataList class}
\description{
This class manages a collection of ErrorData objects.
It can be used to add new ErrorData objects, and to export the collection of errors to a CSV file.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error_data_list}}{A list of ErrorData objects representing the collection of errors.}

\item{\code{user_data}}{The original data frame where the errors were found.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ErrorDataList-new}{\code{ErrorDataList$new()}}
\item \href{#method-ErrorDataList-add_error}{\code{ErrorDataList$add_error()}}
\item \href{#method-ErrorDataList-merge_data_list}{\code{ErrorDataList$merge_data_list()}}
\item \href{#method-ErrorDataList-to_csv}{\code{ErrorDataList$to_csv()}}
\item \href{#method-ErrorDataList-clone}{\code{ErrorDataList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorDataList-new"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorDataList-new}{}}}
\subsection{Method \code{new()}}{
This method is used to initialize the ErrorDataList object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorDataList$new(user_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_data}}{The original data where the errors were found.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorDataList-add_error"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorDataList-add_error}{}}}
\subsection{Method \code{add_error()}}{
This method adds an ErrorData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorDataList$add_error(error_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error_data}}{The ErrorData object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorDataList-merge_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorDataList-merge_data_list}{}}}
\subsection{Method \code{merge_data_list()}}{
This method is used to merge the error data from another ErrorDataList into this one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorDataList$merge_data_list(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{The other ErrorDataList object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorDataList-to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorDataList-to_csv}{}}}
\subsection{Method \code{to_csv()}}{
This method is used to export the errors to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorDataList$to_csv(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The name of the CSV file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorDataList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorDataList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorDataList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
