% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{execute_sql}
\alias{execute_sql}
\title{Execute SQL Queries on a Database}
\usage{
execute_sql(con, sql)
}
\arguments{
\item{con}{Database connection.}

\item{sql}{List of SQL queries to execute.}
}
\description{
Execute SQL Queries on a Database
}
