% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{execute_upgrade_sql_script}
\alias{execute_upgrade_sql_script}
\title{Execute sql}
\usage{
execute_upgrade_sql_script(con, upgrade_script_path)
}
\arguments{
\item{con}{Database connection}

\item{upgrade_script_path}{Upgrade script path.}
}
\description{
Execute sql
}
