% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{extract_sample_from_json}
\alias{extract_sample_from_json}
\title{Extract Actions, File Types, and Sample Types from JSON Content}
\usage{
extract_sample_from_json(json_content)
}
\arguments{
\item{json_content}{A list containing the parsed JSON data.}
}
\value{
A list containing three vectors: actions, file_types, and sample_types.
}
\description{
This function parses a given JSON structure to extract the available actions,
file types, and sample types.
}
\examples{
\dontrun{
  json_path <- "path_to_your_json_file.json"
  json_content <- fromJSON(json_path)
  info <- extract_info_from_json(json_content)
  print(info)
}
}
\keyword{appdata}
