% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{extract_search_criteria}
\alias{extract_search_criteria}
\title{Extract Search Criteria from User CSV File}
\usage{
extract_search_criteria(user_csv, search_type)
}
\arguments{
\item{user_csv}{A path to the user-provided CSV file.}

\item{search_type}{A string indicating the type of search ("barcode" or "study_subject").}
}
\value{
A message indicating whether the required columns were detected.
}
\description{
This function reads a CSV file provided by the user and extracts
search criteria based on a specified search type (either "barcode" or "study_subject").
}
