% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{find_column_name}
\alias{find_column_name}
\title{Find First Matching Column Name}
\usage{
find_column_name(data, potential_names)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{potential_names}{A character vector containing potential column names to search for.}
}
\value{
A string representing the first matching column name, or NA_character_ if none is found.
}
\description{
This function checks for the presence of potential column names within a data frame
and returns the first one that matches, or NA_character_ if none is found.
}
\examples{
df <- data.frame(
  Barcode = 1:5,
  Minus20Freezer = 6:10
)
find_column_name(df, c("Barcode", "Tube ID", "TubeCode"))
find_column_name(df, c("Minus20Freezer", "Minus80Freezer"))

}
