% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{find_header}
\alias{find_header}
\title{Identify the row in a CSV file that matches the header based on required column names.}
\usage{
find_header(user_file, required_user_column_names, valid_header_rows = 1:2)
}
\arguments{
\item{user_file}{A data frame representing the CSV file.}

\item{required_user_column_names}{A character vector of required column names.}

\item{valid_header_rows}{A numeric vector of row numbers to consider as potential header rows.
By default, it considers the first two rows.}
}
\value{
An integer representing the row number where the header was found, or NULL if not found.
}
\description{
Identify the row in a CSV file that matches the header based on required column names.
}
\examples{
\dontrun{
user_data <- read.csv("path_to_user_file.csv", header = FALSE)
find_header(user_data, c("Name", "Age"))
}
}
