% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{format_composition_types}
\alias{format_composition_types}
\title{Format composition types}
\usage{
format_composition_types(n)
}
\arguments{
\item{n}{A numeric vector that represents the number of strains found in a composition}
}
\value{
A character vector of all monoclonal and polyclonal types in the database (1-strain, 2-strain, etc.)
}
\description{
Format composition types
}
