% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV_Utils.R
\name{format_dates}
\alias{format_dates}
\title{format_dates function}
\usage{
format_dates(user_data, parsed_dates, token_mask)

format_dates(user_data, parsed_dates, token_mask)
}
\arguments{
\item{user_data}{The dataframe containing the data to be formatted}

\item{parsed_dates}{The parsed dates obtained from parse_and_validate_dates function}

\item{token_mask}{The token mask obtained from parse_and_validate_dates function}
}
\value{
The formatted user_data

The formatted user_data
}
\description{
This function formats the 'collection_date' column of a dataframe using the parsed_dates and token_mask.

This function formats the 'collection_date' column of a dataframe using the parsed_dates and token_mask.
}
