% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{format_error}
\alias{format_error}
\title{Create Formatting Error Data Frame}
\usage{
format_error(required)
}
\arguments{
\item{required}{A character vector of required column names that were not detected.}
}
\value{
A data frame detailing the formatting errors.
}
\description{
This function creates a data frame detailing the formatting errors for required columns that are missing.
}
\examples{
\dontrun{
format_error(c("Name", "Age"))
}
}
