% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_all_control_types}
\alias{get_all_control_types}
\title{Get all control types}
\usage{
get_all_control_types(control_file = "controls.json", app_file = "app.json")
}
\arguments{
\item{control_file}{A character string indicating the path to the controls.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of control types.
}
\description{
Get all control types
}
\examples{
get_all_control_types()
}
