% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_all_sample_types}
\alias{get_all_sample_types}
\title{Get all sample types}
\usage{
get_all_sample_types(sample_file = "samples.json", app_file = "app.json")
}
\arguments{
\item{sample_file}{A character string indicating the path to the samples.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of sample types.
}
\description{
Get all sample types
}
\examples{
get_all_sample_types()
}
