% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_composition_types}
\alias{get_composition_types}
\title{Get all composition type from the database}
\usage{
get_composition_types(con)
}
\arguments{
\item{con}{A database connection}
}
\value{
A character vector of all monoclonal and polyclonal types in the database
}
\description{
Get all composition type from the database
}
