% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_container_by_control}
\alias{get_container_by_control}
\title{Get containers by specified sample type}
\usage{
get_container_by_control(control_type)
}
\arguments{
\item{control_type}{A character string specifying the type of sample.}
}
\value{
A named list of containers for the specified sample type.
}
\description{
Get containers by specified sample type
}
\examples{
get_container_by_sample(sample_type = "micronix")
}
