% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_container_by_sample}
\alias{get_container_by_sample}
\title{Get containers by specified sample type}
\usage{
get_container_by_sample(
  sample_type,
  sample_file = "samples.json",
  app_file = "app.json"
)
}
\arguments{
\item{sample_type}{A character string specifying the type of sample.}

\item{sample_file}{A character string indicating the path to the samples.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of containers for the specified sample type.
}
\description{
Get containers by specified sample type
}
\examples{
get_container_by_sample(sample_type = "micronix")
}
