% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_container_table_name}
\alias{get_container_table_name}
\title{Get the Relevant Table Name Based on Container Class}
\usage{
get_container_table_name(container_class)
}
\arguments{
\item{container_class}{A string representing the container class.
Currently supports "micronix_tube" and "cryovial_tube".}
}
\value{
A string representing the name of the relevant table.
Will stop and throw an error for invalid or unsupported container classes.
}
\description{
This function returns the name of the relevant database table
based on the provided container class.
}
\examples{
get_container_table_name("micronix_tube") # Should return "micronix_plate"
get_container_table_name("cryovial_tube") # Should return "cryovial_plate"

}
