% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_control_action_types}
\alias{get_control_action_types}
\title{Get action types for a specific control}
\usage{
get_control_action_types(
  control_type,
  control_file = "controls.json",
  app_file = "app.json"
)
}
\arguments{
\item{control_type}{A character string representing the type of the control.}

\item{control_file}{A character string indicating the path to the controls.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of action types or NULL if no actions are defined.
}
\description{
Get action types for a specific control
}
\examples{
get_control_action_types("dbs_sheet")
}
