% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_control_types}
\alias{get_control_types}
\title{Get control types from a JSON file}
\usage{
get_control_types(controls_file = "controls.json")
}
\arguments{
\item{controls_file}{Name of the controls JSON file. Default is "controls.json".}
}
\value{
A named list of control types.
}
\description{
Retrieves a named list for control types with "name" as the names in the list and "type" as the values.
}
