% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_data_file_path}
\alias{get_data_file_path}
\title{Get the path to a data file in the inst/app/www directory}
\usage{
get_data_file_path(file_name)
}
\arguments{
\item{file_name}{The name of the file.}
}
\value{
The path to the file.
}
\description{
This function takes a filename and returns the path to the file in the inst/app/www directory of the package.
}
\examples{
\dontrun{
file_path <- get_data_file_path("mydata.json")
}
}
\keyword{internal}
