% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{get_db_version}
\alias{get_db_version}
\title{Retrieve Current Database Version}
\usage{
get_db_version(db_or_conn = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{db_or_conn}{Either a connection string to the SQLite database or an active connection.}
}
\value{
Character string indicating the current version of the database.
}
\description{
Retrieve Current Database Version
}
