% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_dereferenced_values}
\alias{get_dereferenced_values}
\title{Get dereferenced values using sample_values and shared_values}
\usage{
get_dereferenced_values(dereferenced_keys, sample_values, shared_values)
}
\arguments{
\item{dereferenced_keys}{List of keys to look up.}

\item{sample_values}{List of primary values.}

\item{shared_values}{List of secondary values.}
}
\value{
A flattened list of values.
}
\description{
This function retrieves the values based on the keys provided.
}
