% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_destination_location_by_control}
\alias{get_destination_location_by_control}
\title{Get destination location by specified control type}
\usage{
get_destination_location_by_control(
  control_type,
  control_file = "controls.json",
  app_file = "app.json"
)
}
\arguments{
\item{control_type}{A character string specifying the type of control.}

\item{control_file}{A character string indicating the path to the controls.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of destination locations for the specified control type.
}
\description{
Get destination location by specified control type
}
\examples{
get_destination_location_by_control(control_type = "DBS Sheet")
}
