% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_duplicated_rows}
\alias{get_duplicated_rows}
\title{Find duplicate rows in a data frame}
\usage{
get_duplicated_rows(data)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{A character vector of column names to check for duplicates.}
}
\description{
Find duplicate rows in a data frame
}
