% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{get_environ_file_path}
\alias{get_environ_file_path}
\title{Get environ file path}
\usage{
get_environ_file_path(site_install)
}
\arguments{
\item{site_install}{Logical indicating whether installation is site-wide.}
}
\value{
A character string of the environment file path.
}
\description{
Get environ file path
}
