% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{get_expected_versions}
\alias{get_expected_versions}
\title{Retrieve Expected Versions from JSON File}
\usage{
get_expected_versions(pkgname)
}
\arguments{
\item{pkgname}{The name of the package for which the expected versions
should be retrieved.}
}
\value{
A list containing the expected versions as described in the
'versions.json' file.
}
\description{
This function gets the expected versions of the specified package from
an external 'versions.json' file within the package's 'extdata' directory.
}
\examples{
\dontrun{
get_expected_versions("sampleDB")
}

}
