% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{get_extractions}
\alias{get_extractions}
\title{Get extractions}
\usage{
get_extractions(database = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{database}{A database path}
}
\value{
A tibble with extractions
}
\description{
Get extractions
}
\examples{
database <- Sys.getenv("SDB_PATH")
get_extractions(database)
}
