% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_controls_json}
\alias{get_fields_controls_json}
\title{Load and parse controls JSON}
\usage{
get_fields_controls_json(file_path, references_json_path, control_type, action)
}
\arguments{
\item{file_path}{The path to the JSON file.}

\item{references_json_path}{The path to the references JSON file.}

\item{control_type}{The type of control.}

\item{action}{The action to perform on the control.}
}
\value{
A list with the required and optional fields.
}
\description{
This function loads and parses the controls.json file.
The list can then be manipulated using standard R operations.
}
\examples{
\dontrun{
fields <- get_fields_controls_json("dbs_sheet", "create")
}
}
\keyword{intenal}
