% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_for_action}
\alias{get_fields_for_action}
\title{Get fields for specific action, sample type, and file type}
\usage{
get_fields_for_action(json_data, sample_type, action, file_type)
}
\arguments{
\item{json_data}{A list containing the parsed JSON data.}

\item{sample_type}{The type of the sample.}

\item{action}{The type of the action.}

\item{file_type}{The type of the file.}
}
\value{
A list containing the required, conditional, and optional fields.
}
\description{
This function extracts the required,
conditional, and optional fields for a specific action,
sample type, and file type from the parsed JSON data.
}
\examples{
\dontrun{
json_data <- load_parse_json("path/to/your/file.json")
fields <- get_fields_for_action(json_data, "Micronix", "upload", "na")
}
}
\keyword{internal}
