% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_for_action_control}
\alias{get_fields_for_action_control}
\title{Get fields for specific action, control type, and file type}
\usage{
get_fields_for_action_control(
  json_data,
  references_json_data,
  control_type,
  action
)
}
\arguments{
\item{json_data}{A list containing the parsed JSON data.}

\item{references_json_data}{A list containing the parsed references JSON data.}

\item{control_type}{The type of the control.}

\item{action}{The type of the action.}
}
\value{
A list containing the required, conditional, and optional fields.
}
\description{
This function extracts the required, conditional, and optional fields for a specific action,
control type from the parsed JSON data. It also takes into account exceptions for the given control type.
}
\examples{
\dontrun{
json_data <- load_parse_json("path/to/your/file.json")
fields <- get_fields_for_action_control(json_data, "Micronix", "create")
}
}
\keyword{internal}
