% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_for_action_sample}
\alias{get_fields_for_action_sample}
\title{Get fields for specific action, sample type, and file type}
\usage{
get_fields_for_action_sample(
  json_data,
  references_json_data,
  sample_type,
  action,
  file_type = "na"
)
}
\arguments{
\item{json_data}{A list. The parsed JSON data containing sample details and defaults.}

\item{references_json_data}{A list. Parsed references JSON data with details about locations, containers, etc.}

\item{sample_type}{A character string specifying the type of sample (e.g., "micronix").}

\item{action}{A character string indicating the desired action (e.g., "upload").}

\item{file_type}{A character string defining the type of the file (e.g., "visionmate"). The default is "na" (application standard).}
}
\value{
A list containing fields for the specified sample type, action, and file type.
This includes required, conditional, and optional fields, alongside container and location information.
}
\description{
Extracts the necessary fields for a given action, sample type, and file type from parsed JSON data.
The function considers default configurations and also allows for specific exceptions
based on sample and file types. It has built-in checks to ensure valid input values
for \code{sample_type}, \code{action}, and \code{file_type}.
}
\examples{
\dontrun{
json_data <- load_parse_json("path/to/your/samples.json")
references_data <- load_parse_json("path/to/your/references.json")
fields <- get_fields_for_action_sample(json_data, references_data, "Micronix", "upload", "visionmate")
}
}
\seealso{
\link[jsonlite]{fromJSON} for parsing JSON in R.
}
\keyword{internal}
