% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_from_json}
\alias{get_fields_from_json}
\title{Get fields from JSON file}
\usage{
get_fields_from_json(appdata, ...)
}
\arguments{
\item{appdata}{The name of JSON file. Must be one of "samples", "controls", or "references"}

\item{...}{Additional parameters required by the specific JSON parsing functions}
}
\value{
A list with the required, conditional, and optional fields (for "samples" and "controls") or the fields for the given category (for "references")
}
\description{
This function fetches the fields from a given JSON file based on the file name and additional parameters
}
\keyword{internal}
