% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_fields_samples_json}
\alias{get_fields_samples_json}
\title{Load and parse samples JSON}
\usage{
get_fields_samples_json(
  file_path,
  references_json_path,
  sample_type,
  action,
  file_type
)
}
\arguments{
\item{file_path}{The path to the JSON file.}

\item{references_json_path}{The path to the references JSON file.}

\item{sample_type}{The type of sample.}

\item{action}{The action to perform on the sample.}

\item{file_type}{The type of file.}
}
\value{
A list with the required, conditional, and optional fields.
}
\description{
This function loads and parses the samples.json file,
extracting the fields for a specific sample type, action,
and file type. It also takes into account exceptions for the given sample and file types.
}
\examples{
\dontrun{
fields <- get_fields_samples_json("Micronix", "upload", "na")
}
}
\keyword{internal}
