% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_file_types_for_control}
\alias{get_file_types_for_control}
\title{Get file types for a specific control}
\usage{
get_file_types_for_control(
  control_type,
  control_file = "controls.json",
  app_file = "app.json"
)
}
\arguments{
\item{control_type}{A character string representing the type of the control.}

\item{control_file}{A character string indicating the path to the controls.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of file types or a list with "NA" if not specified.
}
\description{
Get file types for a specific control
}
\examples{
get_file_types_for_control("dbs_sheet")
}
