% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_file_types_for_sample}
\alias{get_file_types_for_sample}
\title{Get file types for a specific sample}
\usage{
get_file_types_for_sample(
  sample_type,
  sample_file = "samples.json",
  app_file = "app.json"
)
}
\arguments{
\item{sample_type}{A character string representing the type of the sample.}

\item{sample_file}{A character string indicating the path to the samples.json file.}

\item{app_file}{A character string indicating the path to the app.json file.}
}
\value{
A named list of file types or a list with "NA" if not specified.
}
\description{
Get file types for a specific sample
}
\examples{
get_file_types_for_sample("micronix")
}
