% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{get_identifiers_from_user_data}
\alias{get_identifiers_from_user_data}
\title{Extract identifiers from user data}
\usage{
get_identifiers_from_user_data(user_data)
}
\arguments{
\item{user_data}{A data frame with at least two columns: 'Strains' and 'Percentages', each containing semicolon-separated values.}
}
\value{
A data frame containing unique identifiers for each row of user data.
}
\description{
This function processes user data to generate a unique identifier based on strains and percentages.
}
