% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{get_matched_studies}
\alias{get_matched_studies}
\title{Retrieve Matched Studies from the Database}
\usage{
get_matched_studies(user_file, database)
}
\arguments{
\item{user_file}{A data frame representing the user's CSV file.}

\item{database}{A character string specifying the path to the SQLite database.}
}
\value{
A data frame containing matched study records.
}
\description{
This function connects to the specified SQLite database, retrieves study records that match
the \code{StudyCode} present in the user's file, and returns these matched studies.
}
\examples{
\dontrun{
user_file <- data.frame(StudyCode = c("ST001", "ST002"))
database_path <- "path_to_database.sqlite"
get_matched_studies(user_file, database_path)
}
}
\keyword{internal}
