% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_normalized_path}
\alias{get_normalized_path}
\title{Get normalized path}
\usage{
get_normalized_path(site_install, pkgname, dir_type, filename)
}
\arguments{
\item{site_install}{Logical indicating whether installation is site-wide.}

\item{pkgname}{Character string of the package name.}

\item{dir_type}{Character string specifying the type of directory. Valid values are "config" and "data".}

\item{filename}{File to add.}
}
\value{
A character string of the normalized path.
}
\description{
Get normalized path
}
