% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{get_package_path}
\alias{get_package_path}
\title{Get Installation Path of a Package}
\usage{
get_package_path(pkgname)
}
\arguments{
\item{pkgname}{Character string specifying the name of the package.}
}
\value{
Character string indicating the installation path of the package,
or \code{NULL} if the package is not found.
}
\description{
Retrieves the installation path of the given package.
}
\examples{
get_package_path("sampleDB")

}
\keyword{internal}
