% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_percentages}
\alias{get_percentages}
\title{Get all batches from the database}
\usage{
get_percentages(con)
}
\arguments{
\item{con}{A database connection}
}
\value{
A character vector of all batches in the database

A character vector of all studies in the database
}
\description{
Get all batches from the database
}
\examples{
con <- dbConnect(RSQLite::SQLite(), ":memory:")
get_batches(con)
dbDisconnect(con)
}
\keyword{database}
\keyword{utility}
