% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_position_column_by_sample}
\alias{get_position_column_by_sample}
\title{Get the expected position column for a sample type}
\usage{
get_position_column_by_sample(
  sample_type,
  sample_file = "samples.json",
  app_file = "app.json"
)
}
\arguments{
\item{sample_type}{A character string specifying the type of sample.}

\item{sample_file}{A character string indicating the path to the samples.json file.}

\item{control_file}{A character string specifying the path to the controls.json file.}
}
\value{
The position column for that sample type.
}
\description{
Micronix have different file formats which report sample positioning in different ways. This
function makes it easy to look up that information.
}
