% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_references_file_columns}
\alias{get_references_file_columns}
\title{Get user action fields for references}
\usage{
get_references_file_columns(reference_type)
}
\arguments{
\item{reference_type}{The reference to get fields for}
}
\value{
A FileColumnAttributes object with fields for the user action for references
}
\description{
This function gets all the fields for a given user action related to references.
}
\keyword{appdata}
