% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_sample_file_columns}
\alias{get_sample_file_columns}
\title{Get user column attributes for samples}
\usage{
get_sample_file_columns(
  sample_type,
  user_action,
  file_type,
  config_yml = Sys.getenv("SDB_CONFIG")
)
}
\arguments{
\item{sample_type}{The type of sample}

\item{user_action}{The user action to get fields for}

\item{file_type}{The type of file}

\item{config_yml}{The path to the config.yml file}
}
\value{
A FileColumnAttributes object with required, conditional, and optional fields for the user action
}
\description{
This function gets all the required, conditional, and optional fields
for a given user action. It also takes into account exceptions for the given sample and file types.
}
\keyword{appdata}
