% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_sample_file_types}
\alias{get_sample_file_types}
\title{Retrieve File Types for Each Sample from JSON}
\usage{
get_sample_file_types(json_path)
}
\arguments{
\item{json_path}{A character string representing the path to the JSON file.}
}
\value{
A named list where each item's name is a sample ID, and the item's value is a named vector
of file types available for that sample, with the name being the capitalized version and the value
being the original file type. Each vector always includes "NA" as one of the options.
}
\description{
This function extracts the available file types for each sample defined in the provided JSON file.
"NA" will always be returned as one of the file types for each sample.
}
\examples{
\dontrun{
file_types <- get_sample_file_types("samples.json")
print(file_types)
}

}
\seealso{
\code{\link{load_parse_json}}
}
