% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_sample_ids_from_json}
\alias{get_sample_ids_from_json}
\title{Extract sample IDs from JSON}
\usage{
get_sample_ids_from_json(file_path)
}
\arguments{
\item{file_path}{The path to the JSON file.}
}
\value{
A vector of sample IDs.
}
\description{
This function loads the provided JSON and returns the sample IDs from the "samples" section.
}
\examples{
\dontrun{
sample_ids <- get_sample_ids_from_json("your_json_file_name.json")
print(sample_ids)
}
}
\keyword{internal}
