% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_sample_types_by_action}
\alias{get_sample_types_by_action}
\title{Get sample types from a JSON file by action}
\usage{
get_sample_types_by_action(action, samples_file = "samples.json")
}
\arguments{
\item{samples_file}{Name of the samples JSON file. Default is "samples.json".}
}
\value{
A named list of sample types.
}
\description{
Retrieves a named list for sample types with "name" as the names in the list and "type" as the values, filtered by action.
}
