% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_unique_compositions_from_database}
\alias{get_unique_compositions_from_database}
\title{Extract unique compositions from the database}
\usage{
get_unique_compositions_from_database(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing sorted_strains_key for each composition in the database with associated strains and percentages.
}
\description{
This function queries a database to collect and process data, generating a unique composition key based on sorted strains and percentages.
}
\seealso{
\code{\link{format_labels}}
}
