% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_unique_compositions_from_user_data}
\alias{get_unique_compositions_from_user_data}
\title{Normalize Percentages by Group}
\usage{
get_unique_compositions_from_user_data(user_data)
}
\arguments{
\item{user_data}{A data frame with at least two columns: 'Strains' and 'Percentages', each containing semicolon-separated values (for polyclonal controls).}
}
\value{
A data frame containing sorted_strains_key for each row of user data.
}
\description{
Normalize percentage column by groups so that
Extract unique compositions from user data
}
\details{
This function processes user data to generate a unique composition key based on sorted strains and percentages.
}
