% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_values}
\alias{get_values}
\title{Get values with priority to sample_values over shared_values.}
\usage{
get_values(keys, sample_values, shared_values)
}
\arguments{
\item{keys}{List of keys for which values are needed.}

\item{sample_values}{List of sample values.}

\item{shared_values}{List of shared values.}
}
\value{
A vector of values prioritized from sample_values, if available.
}
\description{
This function prioritizes and fetches values from either sample_values or shared_values.
}
