% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{handle_special_columns}
\alias{handle_special_columns}
\title{Handle special columns in the user CSV file}
\usage{
handle_special_columns(
  user_file,
  container_name,
  freezer_address,
  location_params
)
}
\arguments{
\item{user_file}{The user's CSV file as a data frame.}

\item{container_name}{The name of the container column if provided by the user.}

\item{freezer_address}{The address of the freezer column if provided by the user.}

\item{location_params}{A character vector of location parameters/columns related to the freezer address.}
}
\value{
A list containing the modified user_file and an updated vector of missing columns.
}
\description{
This function handles special cases for columns like \code{container_name} and \code{freezer_address}
which might be added by users.
}
