% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{handle_unknown_date_tokens}
\alias{handle_unknown_date_tokens}
\title{Handle Unknown Date Tokens}
\usage{
handle_unknown_date_tokens(data, date_col, parsed_dates, token_mask)
}
\arguments{
\item{data}{A data frame.}

\item{date_col}{The name of the date column.}

\item{parsed_dates}{A vector of parsed dates.}

\item{token_mask}{A logical mask identifying rows with recognized tokens indicating unknown dates.}
}
\value{
The adjusted data frame.
}
\description{
Adjusts the date column for rows where the user signified with tokens that the dates are unknown.
}
\keyword{utils}
