% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{handle_upgrade_error}
\alias{handle_upgrade_error}
\title{Handle any errors during the database upgrade}
\usage{
handle_upgrade_error(new_database, e)
}
\arguments{
\item{new_database}{Path to the temporary new database.}

\item{e}{Error message.}
}
\value{
Stops with an error message.
}
\description{
This function provides a way to handle errors that might arise during the database upgrade.
It will clean up any temporary files and disconnect from the database.
}
