% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{init_db_conn}
\alias{init_db_conn}
\title{Initialize a Database Connection}
\usage{
init_db_conn(db_path)
}
\arguments{
\item{db_path}{The path to the SQLite database.}
}
\value{
A connection object to the SQLite database.
}
\description{
This function establishes and returns a connection to an SQLite database.
}
\examples{
\dontrun{
  conn <- init_db_conn(Sys.getenv("SDB_PATH"))
}
}
