% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{initialize_database_with_base_version}
\alias{initialize_database_with_base_version}
\title{Initialize Database with Base Version}
\usage{
initialize_database_with_base_version(
  new_database,
  pkgname,
  base_version = "1.0.0"
)
}
\arguments{
\item{new_database}{Path to the new temporary SQLite database.}

\item{pkgname}{Name of the package containing the database upgrade scripts.}

\item{db_versions}{Vector of available database versions.}
}
\value{
None (primarily called for side effects).
}
\description{
This function initializes the SQLite database using the base version.
}
\examples{
\dontrun{
  initialize_database_with_base_version(tempfile(), "my_package", c("1.0.0", "2.0.0"))
}
}
