% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{is_linux_platform}
\alias{is_linux_platform}
\title{Check if the system platform is Linux.}
\usage{
is_linux_platform()
}
\value{
Logical value indicating if the platform is Linux.
}
\description{
Check if the system platform is Linux.
}
