% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{is_shiny_server_installed}
\alias{is_shiny_server_installed}
\title{Check if Shiny server is installed.}
\usage{
is_shiny_server_installed(shiny_server_path)
}
\arguments{
\item{shiny_server_path}{Path to the Shiny server.}
}
\value{
Logical value indicating if Shiny server is installed.
}
\description{
Check if Shiny server is installed.
}
