% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{is_system_installed}
\alias{is_system_installed}
\title{Check If Package Is Installed System-wide}
\usage{
is_system_installed(pkgname)
}
\arguments{
\item{pkgname}{Character string specifying the name of the package to check.}
}
\value{
Logical value indicating if the package is installed in a system-wide
location (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Determines if the given package is installed in a system-wide library location.
This is assessed based on whether the package's installation path matches any
of the common system library paths (retrieved via \code{.libPaths()}).
}
\examples{
is_system_installed("sampleDB")
is_system_installed("nonexistentPackage")

}
\keyword{internal}
