% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{join_blood_spot_collections}
\alias{join_blood_spot_collections}
\title{Join Blood Spot Collections}
\usage{
join_blood_spot_collections(df.payload, con)
}
\arguments{
\item{df.payload}{A dataframe containing payload data.}

\item{con}{A database connection object.}
}
\value{
A dataframe after joining with blood spot collections.
}
\description{
Join Blood Spot Collections
}
