% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{join_composition_ids}
\alias{join_composition_ids}
\title{Process user data to handle malaria blood control details}
\usage{
join_composition_ids(df.payload, con, composition_id_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{composition_id_col}{Column representing CompositionID.}
}
\value{
A dataframe with processed malaria blood control details.
}
\description{
Process user data to handle malaria blood control details
}
