% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{join_dbs_control_sheets}
\alias{join_dbs_control_sheets}
\title{Join DBS control sheets based on bag ID and label}
\usage{
join_dbs_control_sheets(df.payload, con, sheet_name_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{dbs_bag_id_col}{Column name in df.payload for DBS bag ID.}

\item{label_col}{Column name in df.payload for label.}
}
\value{
A dataframe with joined DBS control sheets.
}
\description{
Join DBS control sheets based on bag ID and label
}
