% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{join_locations_and_bags}
\alias{join_locations_and_bags}
\title{Join locations and bags from user data to the database}
\usage{
join_locations_and_bags(
  df.payload,
  con,
  location_root_col,
  level_I_col,
  level_II_col,
  dbs_bag_col
)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{location_root_col}{Column name in df.payload for the root location.}

\item{level_I_col}{Column name in df.payload for level I location.}

\item{level_II_col}{Column name in df.payload for level II location.}

\item{dbs_bag_col}{Column name in df.payload for DBS bag name.}
}
\value{
A dataframe with joined location and bag details.
}
\description{
Join locations and bags from user data to the database
}
