% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{lock_database}
\alias{lock_database}
\title{Lock a SQLite Database}
\usage{
lock_database(database)
}
\arguments{
\item{database}{Path to the SQLite database to be locked.}
}
\value{
Database connection object after the locking operation.
}
\description{
This function locks an existing SQLite database to prevent further modifications while performing operations.
}
\examples{
\dontrun{
  con <- lock_database("path_to_original.db")
  # Do other operations...
  DBI::dbDisconnect(con)  # Remember to disconnect when done.
}
}
\keyword{internal}
\keyword{setup}
