% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{merge_defaults_with_exceptions}
\alias{merge_defaults_with_exceptions}
\title{Merge Default Values with Exceptions}
\usage{
merge_defaults_with_exceptions(defaults, exceptions)
}
\arguments{
\item{defaults}{A named list of default values.}

\item{exceptions}{A named list of exceptions that will override the defaults.}
}
\value{
A merged list with exceptions overriding the defaults.
}
\description{
This internal function is used to combine a default set of key-value pairs with
an exceptions set, where the exceptions will override the defaults.
}
\examples{
defaults <- list(a = 1, b = 2, c = 3)
exceptions <- list(b = 20, d = 4)
merge_defaults_with_exceptions(defaults, exceptions)

}
\keyword{internal}
