% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{message_not_site_install}
\alias{message_not_site_install}
\title{Display a message indicating the need for a site install.}
\usage{
message_not_site_install()
}
\description{
Display a message indicating the need for a site install.
}
