% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{message_not_supported_platform}
\alias{message_not_supported_platform}
\title{Display a message indicating Shiny server isn't supported.}
\usage{
message_not_supported_platform()
}
\description{
Display a message indicating Shiny server isn't supported.
}
