% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{message_shiny_not_installed}
\alias{message_shiny_not_installed}
\title{Display a message indicating Shiny server isn't installed.}
\usage{
message_shiny_not_installed()
}
\description{
Display a message indicating Shiny server isn't installed.
}
