% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveContainers.R
\name{MoveContainers}
\alias{MoveContainers}
\title{location <- list(name.freezer = "", level_II = "", level_II = "")}
\usage{
MoveContainers(type, container_name, location)
}
\description{
location <- list(name.freezer = "", level_II = "", level_II = "")
}
