% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveSamples.R
\name{MoveSamples}
\alias{MoveSamples}
\title{Move Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
MoveSamples(type, file.container_samples)
}
\arguments{
\item{file.container_samples}{A list containing an item for each plate involved in the move. Key is plate name, value is path to MoveCSV}
}
\description{
Move Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
sampleDB::MoveTubes(file.container_samples = list("container_name_1" = "~/path/container_1_samples.csv", "container_name_2" = "~/path/container_2_samples.csv"))
write.csv(tibble(LocationRow = rep("A", 10), LocationColumn = c(1:10), TubeCode = c(paste("XXXXX", 0:9))), "~/path/container_1_samples.csv")
write.csv(tibble(Position = rep("A", 10), "Tube ID" = c(paste("XXXXX", 0:9)), "Status" = NA,	"Error Count" = NA,	"Rack ID" = NA,	"Date" = NA), "~/path/container_1_samples.csv")
write.csv(tibble(row = rep("A", 10), column = c(1:10), "label" = c(paste("XXXXX", 0:9)), "~/path/container_1_samples.csv")
}
