% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveTubes.R
\name{MoveTubes}
\alias{MoveTubes}
\title{Move Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
MoveTubes(file.barcode)
}
\arguments{
\item{file.barcode}{A list containing an item for each plate involved in the move. Key is plate name, value is path to MoveCSV}
}
\description{
Move Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
sampleDB::MoveTubes(file.barcode = list("name1" = "~/path/name1.csv", "name2" = "~/path/name2.csv"))
}
