% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV_Utils.R
\name{parse_and_validate_dates}
\alias{parse_and_validate_dates}
\title{parse_and_validate_dates function}
\usage{
parse_and_validate_dates(user_data)

parse_and_validate_dates(user_data)
}
\arguments{
\item{user_data}{The dataframe containing the data to be checked}
}
\value{
A list containing parsed_dates, token_mask, and an ErrorData object if any

A list containing parsed_dates, token_mask, and an ErrorData object if any
}
\description{
This function parses and validates dates in the 'collection_date' column of a dataframe. It also checks for tokens representing unknown dates.
Errors found in the date format are returned as an ErrorData object.

This function parses and validates dates in the 'collection_date' column of a dataframe. It also checks for tokens representing unknown dates.
Errors found in the date format are returned as an ErrorData object.
}
