% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{perform_dbs_sample_db_validations}
\alias{perform_dbs_sample_db_validations}
\title{Perform Database-Related Validations for DBS Samples}
\usage{
perform_dbs_sample_db_validations(database, user_data, action)
}
\arguments{
\item{database}{The database connection or specification for validation.}

\item{user_data}{The users data.}

\item{action}{The action being performed, either "upload" or "move".}
}
\value{
A list containing validation errors, if any.
}
\description{
This function conducts database-related validation checks for DBS specimen data.
It initiates a database connection, then performs upload or move validations based on the action provided.
}
\keyword{cryovial}
\keyword{validation,}
