% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{perform_dbs_sheet_db_validations}
\alias{perform_dbs_sheet_db_validations}
\title{Perform Database-Related Validations for DBS Sheets}
\usage{
perform_dbs_sheet_db_validations(database, user_data, action)
}
\arguments{
\item{database}{The database connection or specification for validation.}

\item{user_data}{The users data.}

\item{action}{The action being performed, either "create" or "extraction".}
}
\value{
A list containing validation errors, if any.
}
\description{
This function conducts database-related validation checks for DBS sheet control data.
It initiates a database connection, then performs create or extraction validations based on the action provided.
}
\keyword{dbs_sheet}
\keyword{validation,}
