% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{prepare_new_compositions}
\alias{prepare_new_compositions}
\title{Prepare new compositions for database upload}
\usage{
prepare_new_compositions(user_data)
}
\arguments{
\item{user_data}{A data frame containing user compositions.}
}
\value{
A dataframe with new compositions
}
\description{
Prepare new compositions for database upload
}
