% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{prepare_reference_data_for_validation}
\alias{prepare_reference_data_for_validation}
\title{Prepare DataFrame for Reference Validation}
\usage{
prepare_reference_data_for_validation(
  user_data,
  reference_type,
  file_column_attr
)
}
\description{
Prepare DataFrame for Reference Validation
}
