% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{prepare_specimen_data_for_validation}
\alias{prepare_specimen_data_for_validation}
\title{Prepare DataFrame for Specimen Validation}
\usage{
prepare_specimen_data_for_validation(
  sample_type,
  user_data,
  file_type,
  container_type,
  file_column_attr
)
}
\arguments{
\item{user_data}{A data frame with columns to be renamed.}

\item{file_type}{Parameter needed for internal function. (NOTE: should make this information available in a different way since we need it for file_column_attr)}

\item{file_column_attr}{A ColumnData object containing the column mappings.}
}
\value{
A pre-processed data frame.
}
\description{
This function pre-processes the input data frame based on given column mappings.
It renames columns and also applies column-specific transformations where needed.
}
\examples{
\dontrun{
user_data <- data.frame(
  StudyCode = c(1,2),
  StudySubject = c('X', 'Y'),
  Row = c(1, 2),
  Column = c('A', 'A')
)

# Assuming file_column_attr is an instantiated ColumnData object

new_df <- prepare_specimen_data_for_validation(user_data, file_column_attr)
}
}
