% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{print.ColumnData}
\alias{print.ColumnData}
\title{Print method for ColumnData objects}
\usage{
\method{print}{ColumnData}(x, ...)
}
\arguments{
\item{x}{An object of class 'ColumnData'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This method provides a custom print representation for objects of class 'ColumnData'.
It displays the required, conditional, and optional columns separately.
}
