% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_and_append_compositions}
\alias{process_and_append_compositions}
\title{Process and append new compositions to the database}
\usage{
process_and_append_compositions(con, user_data)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A data frame containing user compositions.}
}
\value{
Integer number of rows appended.
}
\description{
Process and append new compositions to the database
}
