% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_bag_data}
\alias{process_bag_data}
\title{Process user data to handle bag details}
\usage{
process_bag_data(
  df.payload,
  con,
  location_root_col,
  level_I_col,
  level_II_col,
  dbs_bag_name_col
)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{location_root_col}{Column name in df.payload for the root location.}

\item{level_I_col}{Column name in df.payload for level I location.}

\item{level_II_col}{Column name in df.payload for level II location.}

\item{dbs_bag_name_col}{Column name in df.payload for the dbs bag name.}
}
\value{
A dataframe with processed bag details.
}
\description{
Process user data to handle bag details
}
