% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_dbs_control_sheet_data}
\alias{process_dbs_control_sheet_data}
\title{Process user data to handle DBS control sheet details}
\usage{
process_dbs_control_sheet_data(df.payload, con, dbs_sheet_name_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{dbs_sheet_name_col}{Column name in df.payload for DBS Sheet name.}
}
\value{
A dataframe with processed DBS control sheet details.
}
\description{
Process user data to handle DBS control sheet details
}
