% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_malaria_blood_control_data}
\alias{process_malaria_blood_control_data}
\title{Process user data to handle malaria blood control details}
\usage{
process_malaria_blood_control_data(
  df.payload,
  con,
  density_col,
  composition_id_col
)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{density_col}{Column name in df.payload for density value.}

\item{composition_id_col}{Column representing CompositionID.}
}
\value{
A dataframe with processed malaria blood control details.
}
\description{
Process user data to handle malaria blood control details
}
