% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_whole_blood_location_container}
\alias{process_whole_blood_location_container}
\title{Process Whole Blood Box Data}
\usage{
process_whole_blood_location_container(
  user_data,
  con,
  created_col,
  last_updated_col,
  box_name_col,
  box_barcode_col,
  location_root_col,
  level_I_col,
  level_II_col
)
}
\arguments{
\item{user_data}{A dataframe containing the payload data.}

\item{con}{A database connection object.}

\item{created_col}{Column name in user_data for created date.}

\item{last_updated_col}{Column name in user_data for last updated date.}

\item{box_name_col}{Column name in user_data for box name.}

\item{box_barcode_col}{Column name in user_data for box barcode.}
}
\value{
A dataframe with the location ids and cryovial box ids.
}
\description{
This function will check for locations and add new boxes if they don't exist. It will return the boxes and locations IDs.
}
