% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{read_and_preprocess_csv}
\alias{read_and_preprocess_csv}
\title{Read and Preprocess User CSV File}
\usage{
read_and_preprocess_csv(user_csv)
}
\arguments{
\item{user_file}{The path to the user's CSV file. This will be read into a data frame.}
}
\value{
A preprocessed user CSV as a data frame.
}
\description{
This function reads the provided user CSV file and then preprocesses it based on the given user action.
}
\examples{
\dontrun{
user_file_path <- "path/to/user/file.csv"
action <- "upload"

preprocessed_data <- read_and_preprocess_csv(user_file_path, action)
}
}
\seealso{
\code{\link{read_user_csv}}, \code{\link{preprocess_csv}}
}
