% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{rejoin_box_ids}
\alias{rejoin_box_ids}
\title{Rejoin Whole Blood Box Data}
\usage{
rejoin_box_ids(con, user_data, box_name_col, box_barcode_col)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A dataframe containing the payload data.}

\item{box_name_col}{Column name in user_data for box name.}

\item{box_barcode_col}{Column name in user_data for box barcode.}
}
\value{
A dataframe with the location ids and cryovial box ids.
}
\description{
Rejoin Whole Blood Box Data
}
\keyword{internal}
