% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{rejoin_dbs_control_sheet_ids}
\alias{rejoin_dbs_control_sheet_ids}
\title{Rejoin to Get New DBS Control Sheet IDs}
\usage{
rejoin_dbs_control_sheet_ids(df.payload, con, dbs_sheet_name_col)
}
\arguments{
\item{df.payload}{A dataframe containing payload data.}

\item{con}{A database connection object.}
}
\value{
A dataframe after rejoining to get the new DBS control sheet IDs.
}
\description{
Rejoin to Get New DBS Control Sheet IDs
}
