% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{rejoin_malaria_controls}
\alias{rejoin_malaria_controls}
\title{Rejoin malaria controls based on density and study subject ID}
\usage{
rejoin_malaria_controls(df.payload, con, density_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}
}
\value{
A dataframe with joined malaria controls.
}
\description{
Rejoin malaria controls based on density and study subject ID
}
