% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{rejoin_to_get_bag_ids}
\alias{rejoin_to_get_bag_ids}
\title{Rejoin the payload with the database to retrieve bag IDs}
\usage{
rejoin_to_get_bag_ids(df.payload, con, manifest_name_col)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}

\item{manifest_name_col}{Column name in df.payload for manifest name.}
}
\value{
A dataframe with updated bag IDs.
}
\description{
Rejoin the payload with the database to retrieve bag IDs
}
