% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{rejoin_with_updated_blood_spot_collections}
\alias{rejoin_with_updated_blood_spot_collections}
\title{Re-join with updated blood spot collections}
\usage{
rejoin_with_updated_blood_spot_collections(df.payload, con)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}
}
\value{
A dataframe with rejoined blood spot collections.
}
\description{
This function updates the payload dataframe by rejoining with the updated blood spot collections.
It removes the original 'blood_spot_collection_id' from the payload and excludes 'total' and 'exhausted' columns
from the blood spot collections during the join.
}
