% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{remove_na_columns_in_row}
\alias{remove_na_columns_in_row}
\title{Remove columns that have NA in a row}
\usage{
remove_na_columns_in_row(data, row_num)
}
\arguments{
\item{data}{The data frame}

\item{row_num}{The row to look for NAs in}
}
\description{
This is a helper function for set_user_file_header. Before set_header_row
can be called, we need to make sure there are no NAs in the row because
we should not have NAs as column headers. At this point we know what
the header row should be so we can simply remove columns at this point.
This helps resolve bugs - if he user has a random character in a row,
it will create a new column. This potentially could be addressed during
preprocessing but the advantage of handling this in set_user_file_header
is that we have have the columns that we want and we can confidently remove
anything else.
}
