% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{reshape_identifier_to_long_form}
\alias{reshape_identifier_to_long_form}
\title{Reshape Identifier Data to Long Form}
\usage{
reshape_identifier_to_long_form(data)
}
\arguments{
\item{data}{A data frame containing a 'unique_id' column with strain-percentage pairs, separated
by semicolons (';') and hyphens ('-').}
}
\value{
A data frame reshaped to long form, containing new 'strain' and 'percentage' columns.
}
\description{
This function takes a data frame and reshapes it to long form based on the strain-percentage pairs
in the 'unique_id' column. It separates each strain-percentage pair into two new columns: 'strain'
and 'percentage'.
}
