% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{retrieve_compositions_by_identifier}
\alias{retrieve_compositions_by_identifier}
\title{Retrieve compositions from the database by identifier}
\usage{
retrieve_compositions_by_identifier(con, identifiers)
}
\arguments{
\item{con}{A database connection object.}

\item{identifiers}{A character vector of identifiers to be matched.}
}
\value{
A dataframe of matched compositions
}
\description{
This function accepts an identifier that describes the composition of a control, and retrieves data that matches the identifier from the database.
}
