% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{retrieve_compositions_by_label}
\alias{retrieve_compositions_by_label}
\title{Retrieve compositions from the database by label}
\usage{
retrieve_compositions_by_label(con, labels)
}
\arguments{
\item{con}{A database connection object.}

\item{labels}{A character vector of labels to be matched.}
}
\value{
A dataframe of matched compositions
}
\description{
Retrieve compositions from the database by label
}
