% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_SampleDB.R
\name{Run_SampleDB}
\alias{Run_SampleDB}
\title{Run sampleDB as a R Shiny App}
\usage{
Run_SampleDB(host = "0.0.0.0", port = 3838)
}
\description{
Launch sampleDB as a R Shiny application.
}
\examples{
\dontrun{
RunSampleDB()
}
}
