% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely get a value from a list of possible values}
\usage{
safe_extract(data_row, ...)
}
\description{
This function will return the first column that it finds from a list of columns
}
\keyword{internal}
