% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{SampleDB_Setup}
\alias{SampleDB_Setup}
\title{Save SampleDB's database file to a specified location}
\usage{
SampleDB_Setup(env = TRUE, db = TRUE, server = TRUE)
}
\description{
Save SampleDB's database file to a specified location
}
\examples{
\dontrun{
library(sampleDB)
SampleDB_Setup()
}

}
