% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{search_by_study_subject_file_upload}
\alias{search_by_study_subject_file_upload}
\title{Search by Study Subject (and other longitudinal study criteria)}
\usage{
search_by_study_subject_file_upload(
  sample_storage_type,
  filters,
  database = Sys.getenv("SDB_PATH"),
  config_yml = Sys.getenv("SDB_CONFIG")
)
}
\arguments{
\item{sample_storage_type}{A character string indicating the type of sample storage
("micronix", "cryovial").}

\item{filters}{A list of filters to apply to the query.}

\item{database}{An optional parameter indicating the database path. Defaults to the
value of the environment variable 'SDB_PATH'.}

\item{config_yml}{An optional parameter indicating the configuration YAML file. Defaults
to the value of the environment variable 'SDB_CONFIG'.}
}
\value{
A dataframe containing search results based on provided filters and sample storage type.
}
\description{
This function takes a sample storage type, filters, and optional database and config parameters
to search by study subjects and retrieve relevant data.
}
\examples{
\dontrun{
  results <- search_by_study_subject_file_upload("micronix", list(short_code = c("SC01"), study_subject = c("SS01")))
  print(results)
}

}
