% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{search_micronix_tube}
\alias{search_micronix_tube}
\title{Search Micronix Tube Samples}
\usage{
search_micronix_tube(database, micronix_search_df)
}
\arguments{
\item{database}{A string specifying the path to the SQLite database.}

\item{micronix_search_df}{A DataFrame containing the micronix tube samples to be searched.
The DataFrame must contain at least a 'barcode' column.}
}
\value{
A list containing three DataFrames: \code{missing_from_db}, \code{additional_in_db}, and
\code{archived_samples}. Each DataFrame provides details on samples that are missing from the
database, additional in the database but not in the search DataFrame, and samples that are
in an archived status, respectively.
}
\description{
This function searches through Micronix tube samples in the database to find matches
based on a given DataFrame. It identifies discrepancies in positions, missing samples,
or samples that are in an archived status.
}
