% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{SearchControls}
\alias{SearchControls}
\title{Get all malaria blood controls}
\usage{
SearchControls(
  filters,
  control_type = NULL,
  database = Sys.getenv("SDB_PATH"),
  include_internal_control_id = FALSE
)
}
\arguments{
\item{con}{A dplyr dbConnect() connection object}
}
\description{
\code{SearchControls()} can be used to upload controls to the sampleDB database. This function returns lazy-sql.
}
