% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{SearchSamples}
\alias{SearchSamples}
\title{Search for Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
SearchSamples(filters, study_subject.file = FALSE)
}
\arguments{
\item{filters}{A list containing micronix barcodes, plate name, study code, location and/or specimen_type}

\item{study_subject.file}{TRUE or FALSE}
}
\description{
Search for Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
x search.container = #tricky bc are container names uniq throughout the whole db or just within a certain sample type?
- search.date = list(date.from = "YMD", date.to = "YMD")
- search.exhausted = T/F
x search.label = #tricky bc are labels uniq throughout the whole db or just within a certain sample type?
+ search.location = list(location_name = c(), level_I = c(), level_II = c())
+ search.specimen_type = c()
+ search.study = c()
+ search.study_subject = c()
+ search.type = c("RDT", "Micronix", "Paper", "Cryovile")

NOTE: Ufortunately the whole database has to be read into memory inorder to search for something
Im sure there is a faster sql way to do this
Examples:
SearchSamples(filters = list(search.type = c("cryo")))
SearchSamples(filters = list(search.type = c("cryo"), search.label = list(cryovial.labels = c("PPPP 1"))))
SearchSamples(filters = list(search.type = c("cryo", "micronix"), search.label = list(cryovial.labels = c("PPPP 1", "PPPP 2"), micronix.labels = c("XXXX 1"))))
}
