% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{select_relevant_columns}
\alias{select_relevant_columns}
\title{Select relevant columns from the user CSV file}
\usage{
select_relevant_columns(user_file, file_column_attr)
}
\arguments{
\item{user_file}{The user's CSV file as a data frame.}

\item{file_column_attr}{An instance of ColumnData containing attributes of file columns.}
}
\value{
A data frame with only the relevant columns selected.
}
\description{
This function selects the required, conditional, and optional columns from the user file
based on provided file column attributes.
}
