% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{set_header_row}
\alias{set_header_row}
\title{Set a Specific Row as the Header}
\usage{
set_header_row(user_file, header_row)
}
\arguments{
\item{user_file}{A data frame representing the user file.}

\item{header_row}{A numeric value indicating which row to set as the header.}
}
\value{
A data frame with the updated header.
}
\description{
This function updates the header of the user file based on the specified row number.
}
\examples{
\dontrun{
user_file <- read.csv("path_to_user_file.csv", header = FALSE)
set_header_row(user_file, 2)
}
}
