% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{setup_database}
\alias{setup_database}
\title{Setup or upgrade the database}
\usage{
setup_database(expected_database_version, pkgname, database)
}
\arguments{
\item{expected_database_version}{A string that contains the expected database version for the application release.}

\item{pkgname}{Name of the R package.}

\item{database}{Path to the SQLite database file.}
}
\value{
NULL (The function is primarily called for its side effects).
}
\description{
This function sets up a new SQLite database or upgrades it if it already exists.
The function will:
\enumerate{
\item Check the current version of the database.
\item Create a new one if none exists or upgrades an existing database to the target version.
\item Handle errors during the setup or upgrade process and revert any changes.
}
}
\examples{
\dontrun{
  setup_database(list(database = "2.0.0"))
}
}
