% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{setup_directory}
\alias{setup_directory}
\title{Setup a directory if it does not exist.}
\usage{
setup_directory(directory)
}
\arguments{
\item{directory}{The path to the directory to be created.}
}
\value{
Returns TRUE if the directory is created, FALSE otherwise.
}
\description{
Setup a directory if it does not exist.
}
\examples{
setup_directory("new_directory_path")

}
