% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{setup_environment}
\alias{setup_environment}
\title{Main environment setup function}
\usage{
setup_environment(site_install, pkgname, expected_versions, database)
}
\arguments{
\item{site_install}{Logical indicating whether installation is site-wide.}

\item{pkgname}{Character string of the package name.}

\item{expected_versions}{Named list of expected versions for components.}

\item{database}{The path to the database.}
}
\value{
None.
}
\description{
Setups the environment based on the provided settings.
This includes setting up the configuration, database location and subdirectories.
}
