% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{split_and_sort}
\alias{split_and_sort}
\title{Split and sort strains and percentages}
\usage{
split_and_sort(strains, percentages)
}
\arguments{
\item{strains}{A character string with strains separated by semicolons.}

\item{percentages}{A character string with percentages separated by semicolons.}
}
\value{
A list containing sorted strains and sorted percentages.
}
\description{
This function splits strains and percentages strings and sorts them based on percentages.
}
