% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{split_and_unnest_columns}
\alias{split_and_unnest_columns}
\title{Split and Unnest Strain and Percentage Columns}
\usage{
split_and_unnest_columns(user_data, strains_col, percentage_col, append = NULL)
}
\arguments{
\item{user_data}{A data frame with the specified strain and percentage columns.}

\item{strains_col}{The column name in the data frame containing strain data.}

\item{percentage_col}{The column name in the data frame containing percentage data.}
}
\value{
A data frame with unnested strain and percentage columns.
}
\description{
This function takes in a data frame and splits the provided strain and percentage columns by a semicolon (;).
The resulting two lists within each row are then unnested into separate rows. This is commonly used for
validating strains and their corresponding compositions.
}
\examples{
\dontrun{
result <- split_and_unnest_columns(user_data, "strain_column_name", "percentage_column_name")
}

}
\keyword{compositions}
\keyword{pre-validation}
\keyword{strains}
