% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{split_and_unnest_data}
\alias{split_and_unnest_data}
\title{Split and unnest the strain and percentage columns}
\usage{
split_and_unnest_data(data)
}
\arguments{
\item{data}{A data frame with columns named 'strain' and 'percentage'.}
}
\value{
A list with the processed data and potential ErrorData.
}
\description{
Split and unnest the strain and percentage columns
}
\keyword{compositions}
