% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{stop_app_error}
\alias{stop_app_error}
\title{Stop with an application error}
\usage{
stop_app_error(type, message, data = NULL)
}
\arguments{
\item{type}{The type of the error. This should be a character string.}

\item{message}{The error message. This should be a character string.}

\item{data}{Additional data associated with the error.}
}
\value{
Never returns, because it signals an error.
}
\description{
This function throws an application error of the given type,
with the provided message and any additional data.
}
\examples{
\dontrun{
stop_app_error("my_error", "This is an error!", data.frame(x = 1:5))
}

}
