% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{stop_validation_error}
\alias{stop_validation_error}
\title{Stop with a validation error}
\usage{
stop_validation_error(message, data, call = rlang::caller_env())
}
\arguments{
\item{message}{The error message. This should be a character string.}

\item{data}{Additional data associated with the error.}
}
\value{
Never returns, because it signals an error.
}
\description{
This function throws a validation error with the provided message and data.
}
\examples{
\dontrun{
stop_formatting_error("This is a validation error!", data.frame(x = 1:5))
}

}
\keyword{error-handling}
