% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{update_blood_spot_collection_dbs_control_sheet}
\alias{update_blood_spot_collection_dbs_control_sheet}
\title{Update the junction table for blood spot collections and DBS control sheets}
\usage{
update_blood_spot_collection_dbs_control_sheet(df.payload, con)
}
\arguments{
\item{df.payload}{A dataframe with payload data.}

\item{con}{A database connection object.}
}
\value{
A dataframe after updating the junction table.
}
\description{
Update the junction table for blood spot collections and DBS control sheets
}
