% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{update_configuration_file}
\alias{update_configuration_file}
\title{Update Configuration File}
\usage{
update_configuration_file(pkgname, expected_config_version)
}
\arguments{
\item{pkgname}{The name of the package containing the new configuration.}

\item{expected_config_version}{A string with the expected user configuration version.}
}
\value{
NULL (The function is primarily called for its side effects).
}
\description{
This function updates the configuration file based on a package's new configuration.
If the current configuration is missing or outdated, it will be merged with or replaced by the new configuration.
}
\examples{
\dontrun{
update_configuration_file("my_package_name", "1.1.0")
}
}
