% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{update_env_variable}
\alias{update_env_variable}
\title{Update environment variable}
\usage{
update_env_variable(name, value, environ_file_path)
}
\arguments{
\item{name}{Character string of the environment variable name.}

\item{value}{Character string of the environment variable value.}

\item{environ_file_path}{Character string of the path to the environment file.}
}
\value{
None.
}
\description{
Update environment variable
}
