% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDB_Setup.R
\name{upgrade_database}
\alias{upgrade_database}
\title{Upgrade Database}
\usage{
upgrade_database(
  database,
  current_version,
  expected_version,
  db_versions,
  pkgname
)
}
\arguments{
\item{database}{Path to database toupgrade}

\item{current_version}{Current version of the database.}

\item{expected_version}{Expected version of the database.}

\item{db_versions}{Vector of available database versions.}

\item{pkgname}{Name of package}
}
\description{
Upgrade Database
}
