% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{upload_compositions}
\alias{upload_compositions}
\title{Upload compositions to the database}
\usage{
upload_compositions(user_data, database = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{user_data}{A data frame containing user compositions.}

\item{database}{A character string representing the database path. Default is a system environment variable "SDB_PATH".}
}
\value{
A formatted list of labels representing the added or matched compositions.
}
\description{
This function uploads user data to the database after generating unique identifiers and matching them with existing entries.
}
