% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{upload_controls}
\alias{upload_controls}
\title{Upload Controls to the Database}
\usage{
upload_controls(user_data, control_type, database = Sys.getenv("SDB_PATH"))
}
\arguments{
\item{user_data}{Data to be uploaded. It should match the expected format for the specified control type.}

\item{control_type}{The type of control to be uploaded. Valid values are 'dbs_sheet' and 'whole_blood'.}

\item{database}{The path to the database. Defaults to the system environment variable "SDB_PATH".}
}
\value{
NULL. However, the function provides messages to inform the user about the number of uploaded controls or any encountered errors.
}
\description{
This function uploads control data to the specified database based on the control type.
Currently supports uploading for 'dbs_sheet' and 'whole_blood' control types.
}
\examples{
# Assuming you have a data frame called df_dbs_sheet and df_whole_blood:
upload_controls(df_dbs_sheet, control_type = "dbs_sheet")
upload_controls(df_whole_blood, control_type = "whole_blood")

}
