% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSpecimens.R
\name{upload_extracted_dna}
\alias{upload_extracted_dna}
\title{Upload Extracted DNA Data to SampleDB}
\usage{
upload_extracted_dna(
  user_data,
  control_extraction,
  database = Sys.getenv("SDB_PATH")
)
}
\arguments{
\item{user_data}{A data frame containing the DNA extraction data.}

\item{control_extraction}{A character string specifying the type of control extraction.}

\item{database}{A character string specifying the path to the SampleDB SQLite database.
Default value is the system environment variable "SDB_PATH".}
}
\value{
A message indicating the success of the data upload operation.
}
\description{
This function uploads extracted DNA data from a user-provided data frame
to a SQLite database, specifically to SampleDB. The function also updates the
active control count and commits the changes to the database.
}
\examples{
\dontrun{
upload_extracted_dna(user_data = my_data_frame, control_extraction = "dbs_sheet")
}
}
\seealso{
\code{\link{.UpdateActiveControlCount}} and \code{\link{.UploadSpecimens}} for helper functions.
}
