% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSpecimens.R
\name{UploadExtractedDNA}
\alias{UploadExtractedDNA}
\title{Upload Extracted DNA Controls}
\usage{
UploadExtractedDNA(
  user_data,
  control_extraction,
  database = Sys.getenv("SDB_PATH")
)
}
\arguments{
\item{user_data}{A dataframe of SampleDB Upload data.}

\item{control_extraction}{How the DNA was extracted. Should be "DNA (DBS)" or "DNA (WB)"}

\item{con}{A dplyr dbConnect() connection object}
}
\description{
\code{UploadExtractedDNA()} can be used to upload extracted dbs dna to the sampleDB database.
}
