% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{UploadWholeBlood}
\alias{UploadWholeBlood}
\title{Upload DBS Sheet Controls}
\usage{
UploadWholeBlood(user_data, database)
}
\arguments{
\item{user_data}{A dataframe of SampleDB Upload data.}

\item{con}{A dplyr dbConnect() connection object}
}
\description{
\code{UploadDBSSheet()} can be used to upload controls to the sampleDB database.
}
