% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV_Utils.R
\name{validate_and_copy_dates}
\alias{validate_and_copy_dates}
\title{validate_and_copy_dates function}
\usage{
validate_and_copy_dates(user_data, user_action)

validate_and_copy_dates(user_data, user_action)
}
\arguments{
\item{user_data}{The dataframe containing the data to be checked}

\item{user_action}{The action performed by the user, currently only 'upload' is supported}
}
\value{
A list containing user_data and an ErrorDataList object if any

A list containing user_data and an ErrorDataList object if any
}
\description{
This function serves as a pipeline that parses, validates, and formats dates in the 'collection_date' column of a dataframe.
It checks for tokens representing unknown dates. Errors found in the date format are returned as an ErrorDataList object.

This function serves as a pipeline that parses, validates, and formats dates in the 'collection_date' column of a dataframe.
It checks for tokens representing unknown dates. Errors found in the date format are returned as an ErrorDataList object.
}
