% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_box_uniqueness}
\alias{validate_box_uniqueness}
\title{Check if boxes have enough unique cryovial barcodes to be easily identifiable.}
\usage{
validate_box_uniqueness(
  con,
  user_data,
  row_number_col,
  cryovial_col,
  cryovial_box_col,
  similarity_tolerance = 10
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the table where the user data is temporarily stored in the database.}

\item{row_number_col}{The column with the row number in the \code{user_data}.}

\item{cryovial_col}{The column with Cryovial barcodes in the \code{user_data}.}

\item{cryovial_box_col}{The column with Cryovial box IDs in the \code{user_data}.}

\item{similarity_tolerance}{The percent similarity that is tolerated.}
}
\value{
An instance of the ErrorData class or NULL.
}
\description{
This function is meant to check whether a box of cryovials is overly similar to another
box in that it could potentially be confused to the end user. This can also be used to quickly
check if a cryovial box has already been uploaded to the database (ie. prevent duplicated uploads).
This is a problem for cryovials because barcodes must only be unique within a box; they are not universally unique.
}
\keyword{cryovial}
\keyword{validation,}
