% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_composition_sum}
\alias{validate_composition_sum}
\title{Validate if control compositions sum correctly}
\usage{
validate_composition_sum(
  user_data,
  row_number_col,
  strains_col,
  percentage_col,
  tolerance = 2
)
}
\arguments{
\item{user_data}{A data frame containing the user data.}

\item{row_number_col}{The column name for row numbers in user_data.}

\item{percentage_col}{The column name for percentage in user_data.}

\item{tolerance}{Tolerance for sum of control compositions.}
}
\value{
An instance of the ErrorData class or NULL.
}
\description{
Validate if control compositions sum correctly
}
\keyword{validation}
