% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_compositions}
\alias{validate_compositions}
\title{Validate Compositions Data}
\usage{
validate_compositions(con, user_data)
}
\arguments{
\item{con}{The database connection.}

\item{user_data}{A data frame containing composition data to validate.}
}
\value{
An ErrorDataList object containing validation errors, if any.
}
\description{
This function conducts a series of validation checks on composition data.
The function will return any errors encountered during validation.
}
\keyword{validation}
