% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_control_uid_in_batch}
\alias{validate_control_uid_in_batch}
\title{Validate Control UID exists within a batch}
\usage{
validate_control_uid_in_batch(
  con,
  user_data,
  row_number_col,
  control_uid_col,
  batch_col,
  error_if_exists
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A dataframe containing the new extractions to be validated.
It must contain columns that can be used to derive blood_spot_collection_id indirectly.}

\item{row_number_col}{The name of the column in \code{user_data} that provides a unique row identifier.}

\item{control_uid_col}{The name of the column in \code{user_data} that corresponds to the control UID.}

\item{batch_col}{The name of the column in \code{user_data} related to the batch information.}
}
\value{
An \code{ErrorData} object if any blood spot collection's new total extractions exceed its total counts,
detailing the violations. Returns \code{NULL} if all new extraction counts are within the allowed limits.
}
\description{
Validate that a control uid exists in the batch
}
