% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_controls}
\alias{validate_controls}
\title{Validate Control Specimens}
\usage{
validate_controls(database, user_data, control_type, user_action)
}
\arguments{
\item{database}{A database connection or reference.}

\item{control_type}{The type of control specimen ('dbs_sheet' or 'whole_blood').}

\item{user_action}{The user action taken on the control specimen ('extraction' or 'create').}

\item{formatted_csv}{The CSV data formatted for validation.}
}
\value{
Returns the validated user data if no errors are found; stops with validation errors otherwise.
}
\description{
Validate Control Specimens
}
