% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_cryovial_collection_dates}
\alias{validate_cryovial_collection_dates}
\title{Confirm that Cryovials with no barcodes have collection dates if there's already a sample from that study subject.}
\usage{
validate_cryovial_collection_dates(
  con,
  table_name,
  row_number_col,
  study_short_code_col,
  study_subject_col,
  barcode_col,
  collection_date_col
)
}
\arguments{
\item{con}{The database connection.}

\item{table_name}{The table name in the database that contains the data.}

\item{row_number_col}{The column name for row numbers.}

\item{study_short_code_col}{The column name for study short codes.}

\item{study_subject_col}{The column name for study subjects.}

\item{barcode_col}{The column name for barcodes.}

\item{collection_date_col}{The column name for collection dates.}
}
\value{
ErrorData object or NULL if no errors found.
}
\description{
Confirm that Cryovials with no barcodes have collection dates if there's already a sample from that study subject.
}
\keyword{dates}
\keyword{validation,}
