% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_cryovial_moves}
\alias{validate_cryovial_moves}
\title{Validate Cryovial Moves}
\usage{
validate_cryovial_moves(cryovial_test)
}
\arguments{
\item{con}{The database connection.}

\item{cryovial_upload_test}{The utility function for performing tests.}
}
\value{
A list object containing validation errors, if any.
}
\description{
Conducts specific validation checks for moving Cryovial specimens.
}
\keyword{validation}
