% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_cryovial_position}
\alias{validate_cryovial_position}
\title{Validate Cryovial-specific position rules}
\usage{
validate_cryovial_position(data, position_column, row_number_column)
}
\arguments{
\item{data}{A dataframe containing the data to be validated.}

\item{position_column}{A character string specifying the column name for 'position'.}

\item{row_number_column}{A character string specifying the column name for 'RowNumber'.}
}
\value{
An ErrorData object if Cryovial position rules are violated, or NULL otherwise.
}
\description{
Validate Cryovial-specific position rules
}
