% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_cryovial_uploads}
\alias{validate_cryovial_uploads}
\title{Validate Cryovial Uploads}
\usage{
validate_cryovial_uploads(cryovial_test)
}
\arguments{
\item{cryovial_test}{The utility function for performing tests.}

\item{con}{The database connection.}
}
\value{
A list object containing validation errors, if any.
}
\description{
Conducts specific validation checks for uploading Cryovial specimens.
}
\keyword{validation}
