% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_date_format}
\alias{validate_date_format}
\title{Validate Date Format}
\usage{
validate_date_format(data, date_col, allowed_date_formats)
}
\arguments{
\item{data}{A data frame.}

\item{date_col}{The name of the date column.}

\item{allowed_date_formats}{A character vector of allowed date formats.}
}
\value{
An ErrorData object or NULL if no errors.
}
\description{
Validates the date column in the provided data frame and ensures the dates are in a recognized format.
}
