% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_dates_with_tokens}
\alias{validate_dates_with_tokens}
\title{Validate Dates with Tokens}
\usage{
validate_dates_with_tokens(data, date_col, allowed_date_formats, tokens)
}
\arguments{
\item{data}{A data frame.}

\item{date_col}{The name of the date column.}

\item{allowed_date_formats}{A character vector of allowed date formats.}

\item{tokens}{Recognized tokens indicating unknown dates.}
}
\value{
A list with elements \code{error_data} (an ErrorData object or NULL if no errors), \code{parsed_dates} (a vector of parsed dates), and \code{token_mask} (a logical mask identifying rows with recognized tokens indicating unknown dates).
}
\description{
Validates the date column in the provided data frame. It ensures the dates are in a recognized format or a set of allowed tokens indicating the date is unknown.
}
