% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_dbs_bag_label_is_unique}
\alias{validate_dbs_bag_label_is_unique}
\title{Validate DBS Bag Label is unique at a location}
\usage{
validate_dbs_bag_label_is_unique(
  con,
  table_name,
  row_number_col,
  bag_label_col,
  freezer_col,
  shelf_col,
  basket_col,
  error_if_exists = FALSE
)
}
\arguments{
\item{con}{A database connection.}

\item{table_name}{The name of the table to check.}

\item{row_number_col}{The name of the row number column.}

\item{bag_label_col}{The name of the bag label column.}

\item{freezer_col}{The name of the freezer column.}

\item{shelf_col}{The name of the shelf column.}

\item{basket_col}{The name of the basket column.}

\item{error_if_exists}{If TRUE, returns an error if the bag label exists. If FALSE,
returns an error if the bag label does not exist.}
}
\value{
A list containing validation errors, if any.
}
\description{
This function checks that the DBS bag label is unique at a location.
}
\keyword{validation}
