% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_dbs_box_bag_exists}
\alias{validate_dbs_box_bag_exists}
\title{Check DBS Specimen Box and Bag exist}
\usage{
validate_dbs_box_bag_exists(
  con,
  user_data,
  row_number_col,
  container_name_col,
  container_type_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The column name representing the row number in \code{table_name}.}

\item{container_name_col}{The column name representing the Cryovial container name in \code{table_name}.}

\item{label_col}{The column name representing the position in \code{table_name}.}
}
\value{
An object of class ErrorData. If there are no errors, NULL is returned.
}
\description{
This function checks if the Box or Bag exists in the database.
}
\keyword{dbs-specimen}
\keyword{validation,}
