% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_dbs_sample}
\alias{validate_dbs_sample}
\title{Validate DBS Specimen Data}
\usage{
validate_dbs_sample(user_data, action, database)
}
\arguments{
\item{user_data}{A data frame containing specimen data to validate.}

\item{action}{The action being performed, e.g. "upload" or "move".}

\item{database}{The database connection or specification to use for validation.}
}
\value{
A list containing validation errors, if any.
}
\description{
This function conducts a series of validation checks on DBS specimen data.
The function will return any errors encountered during validation.
}
\keyword{cryovial}
\keyword{validation,}
