% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_dbs_sample_label_uniqueness}
\alias{validate_dbs_sample_label_uniqueness}
\title{DBS Specimen paper identifier uniqueness check}
\usage{
validate_dbs_sample_label_uniqueness(
  con,
  user_data,
  row_number_col,
  label_col,
  container_name_col,
  container_type_col,
  error_if_exists
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The column name representing the row number in \code{table_name}.}

\item{label_col}{The column name representing the position in \code{table_name}.}

\item{container_name_col}{The column name representing the Cryovial container name in \code{table_name}.}
}
\value{
An object of class ErrorData. If there are no errors, NULL is returned.
}
\description{
This function checks if the provided paper in the DBS specimen dataset is unique by position.
}
\keyword{dbs-specimen}
\keyword{validation,}
