% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_dbs_sample_moves}
\alias{validate_dbs_sample_moves}
\title{Validate DBS Specimen Moves}
\usage{
validate_dbs_sample_moves(cryovial_test, variable_colnames)
}
\arguments{
\item{con}{The database connection.}

\item{cryovial_upload_test}{The utility function for performing tests.}
}
\value{
A list object containing validation errors, if any.
}
\description{
Conducts specific validation checks for moving DBS Specimen specimens.
}
\keyword{validation}
