% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_dbs_sample_uploads}
\alias{validate_dbs_sample_uploads}
\title{Validate Micronix Uploads}
\usage{
validate_dbs_sample_uploads(dbs_sample_test)
}
\arguments{
\item{micronix_test}{The utility function for performing validation checks.}
}
\value{
A list containing validation errors, if any.
}
\description{
Conducts specific validation checks for uploading Micronix specimens.
}
\keyword{validation}
