% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_dbs_sheet}
\alias{validate_dbs_sheet}
\title{Validate DBS Sheet Control Data}
\usage{
validate_dbs_sheet(user_data, action, database)
}
\arguments{
\item{user_data}{A data frame containing control data to validate.}

\item{action}{The action being performed, either "create" or "extraction".}

\item{database}{The database connection or specification to use for validation.}
}
\value{
A list containing validation errors, if any.
}
\description{
This function conducts a series of validation checks on DBS sheet control data.
The function will return any errors encountered during validation.
}
\keyword{dbs_sheet}
\keyword{validation,}
