% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_dbs_sheet_extraction}
\alias{validate_dbs_sheet_extraction}
\title{Validate DBS Sheet Extraction}
\usage{
validate_dbs_sheet_extraction(dbs_sheet_test)
}
\arguments{
\item{dbs_sheet_test}{The utility function for performing validation checks.}
}
\value{
A list containing validation errors, if any.
}
\description{
Conducts specific validation checks for extracting DBS sheet controls.
}
\keyword{validation}
