% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_dbs_sheet_label_maps_to_uids}
\alias{validate_dbs_sheet_label_maps_to_uids}
\title{Validate that a sheet name for a batch consistently contains
the same densities and compositions. Sheet names (or labels)
should be descriptive of the contents that are on the sheet. This
validation check is in place to enforce sheet name specificity
by only allowing a sheet name to link to composition types and
densities that we're used with the first named sheet on upload.}
\usage{
validate_dbs_sheet_label_maps_to_uids(
  con,
  table_name,
  row_number_col,
  batch_col,
  composition_id_col,
  density_col,
  sheet_name_col
)
}
\arguments{
\item{con}{A database connection.}

\item{table_name}{The name of the table to check.}

\item{row_number_col}{The name of the row number column.}

\item{sheet_name_col}{The name of the freezer column.}

\item{control_uid_col}{The name of the bag label column.}
}
\value{
A list containing validation errors, if any.
}
\description{
New sheets are allowed to reuse the sheet name label but they
must have the same composition type and density combinations.
}
\keyword{validation}
