% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_empty_micronix_well_upload}
\alias{validate_empty_micronix_well_upload}
\title{Micronix Empty Well Validation}
\usage{
validate_empty_micronix_well_upload(
  con,
  table_name,
  row_number_col,
  position_col,
  container_name_col,
  container_barcode_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The column name representing the row number in \code{table_name}.}

\item{position_col}{The column name representing the position in \code{table_name}.}

\item{container_name_col}{The column name representing the Micronix container name in \code{table_name}.}

\item{container_barcode}{col The column name representing the Micronix container barcode in \code{table_name}.}
}
\value{
An object of class ErrorData. If there are no errors, NULL is returned.
}
\description{
This function checks if the provided sample in the Micronix dataset is being uploaded to an empty well.
}
\keyword{micronix}
\keyword{validation,}
