% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_existing_barcodes_by_study}
\alias{validate_existing_barcodes_by_study}
\title{Validate if barcodes already exist for a given study}
\usage{
validate_existing_barcodes_by_study(
  con,
  user_data,
  row_number_col,
  barcode_col,
  study_short_code_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{A data frame containing the user data.}

\item{row_number_col}{The name of the column containing the row number.}

\item{barcode_col}{The column name for barcodes in user_data.}

\item{study_short_code_col}{The column name for study short codes in user_data.}
}
\value{
An instance of the ErrorData class or NULL.
}
\description{
Validate if barcodes already exist for a given study
}
\keyword{cryovial}
\keyword{validation,}
