% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_location_reference_db}
\alias{validate_location_reference_db}
\title{Validate location references (Database Version)}
\usage{
validate_location_reference_db(
  con,
  table_name,
  row_number_col,
  name_col,
  level_I_col,
  level_II_col
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The name of the row number column.}

\item{name_col}{The name of the name column for locations.}

\item{level_I_col}{The name of the Level I column for locations.}

\item{level_II_col}{The name of the Level II column for locations.}
}
\value{
ErrorData object indicating any locations not found.
}
\description{
This function checks if the location reference provided in the dataset exists in the database.
Note: This function assumes that the data is already present in the database.
}
\keyword{validation}
