% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_longitudinal_study}
\alias{validate_longitudinal_study}
\title{Ensure uniqueness of study subjects and collection dates within longitudinal studies.}
\usage{
validate_longitudinal_study(
  con,
  table_name,
  row_number_col,
  study_short_code_col,
  study_subject_col,
  collection_date_col
)
}
\arguments{
\item{con}{The database connection.}

\item{table_name}{The table name in the database that contains the data.}

\item{row_number_col}{The column name for row numbers.}

\item{study_short_code_col}{The column name for study short codes.}

\item{study_subject_col}{The column name for study subjects.}

\item{collection_date_col}{The column name for collection dates.}
}
\value{
ErrorData object or NULL if no errors found.
}
\description{
Ensure uniqueness of study subjects and collection dates within longitudinal studies.
}
\keyword{studies}
\keyword{validation,}
