% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_matrix_container}
\alias{validate_matrix_container}
\title{Validate Matrix Container position}
\usage{
validate_matrix_container(
  con,
  user_data,
  row_number_col,
  container_name_col,
  container_position_col,
  matrix_tablename,
  container_tablename,
  error_if_exists
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the table where the user data is temporarily stored in the database.}

\item{row_number_col}{The column with the row number in the \code{user_data}.}

\item{container_name_col}{The column with container name in \code{user_data}.}

\item{container_position_col}{The column with container positions in the \code{user_data}.}

\item{matrix_tablename}{The tablename in the database that should be checked.}

\item{container_tablename}{The sample container tablename. This is used for position checks.}

\item{error_if_exists}{Logical. If TRUE, an error is returned if the barcode exists in the database.}
}
\value{
An ErrorData object if Micronix position rules are violated, or NULL otherwise.
}
\description{
Validate Matrix Container position
}
