% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_micronix}
\alias{validate_micronix}
\title{Validate Micronix Specimen Data}
\usage{
validate_micronix(user_data, action, database)
}
\arguments{
\item{user_data}{A data frame containing specimen data to validate.}

\item{action}{The action being performed, e.g. "upload" or "move".}

\item{database}{The database connection or specification to use for validation.}
}
\value{
An ErrorDataList object containing validation errors, if any.
}
\description{
This function conducts a series of validation checks on Micronix specimen data.
The function will return any errors encountered during validation.
}
\keyword{micronix}
\keyword{validation,}
