% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_micronix_barcode_length}
\alias{validate_micronix_barcode_length}
\title{Validate Micronix-specific barcode length rules}
\usage{
validate_micronix_barcode_length(data, barcode_column, row_number_column)
}
\arguments{
\item{data}{A dataframe containing the data to be validated.}

\item{barcode_column}{A character string specifying the column name for 'barcode'.}

\item{row_number_column}{A character string specifying the column name for 'RowNumber'.}
}
\value{
An ErrorData object if Micronix barcode rules are violated, or NULL otherwise.
}
\description{
Validate Micronix-specific barcode length rules
}
