% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_non_longitudinal_study_subjects}
\alias{validate_non_longitudinal_study_subjects}
\title{Check uniqueness of StudySubject within non-longitudinal studies.}
\usage{
validate_non_longitudinal_study_subjects(
  con,
  table_name,
  row_number_col,
  study_short_code_col,
  study_subject_col,
  specimen_type_col
)
}
\arguments{
\item{con}{The database connection.}

\item{table_name}{The table name in the database that contains the data.}

\item{row_number_col}{The column name for row numbers.}

\item{study_short_code_col}{The column name for study short codes.}

\item{study_subject_col}{The column name for study subjects.}

\item{specimen_type_col}{The column name for the speciment type.}
}
\value{
ErrorData object or NULL if no errors found.
}
\description{
Check uniqueness of StudySubject within non-longitudinal studies.
}
\keyword{cryovial}
\keyword{validation,}
