% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_references}
\alias{validate_references}
\title{Validate References}
\usage{
validate_references(database, formatted_csv, reference_type, user_action)
}
\arguments{
\item{database}{The database connection or specification for validation.}

\item{formatted_csv}{The user-provided CSV file to be validated.}

\item{reference_type}{The type of reference being validated: "strains" or "compositions".}

\item{user_action}{The action being performed, e.g. "upload".}
}
\value{
The user data if validation passes or an error if not.
}
\description{
The main function to validate references. It can handle different types of references.
After validation, it will either return the user data or stop the process if errors are found.
}
\keyword{validation}
