% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_specimens}
\alias{validate_specimens}
\title{Validate Specimens Main Function}
\usage{
validate_specimens(
  user_data,
  sample_type,
  user_action,
  file_type,
  database = Sys.getenv("SDB_PATH")
)
}
\arguments{
\item{user_data}{Character string. The path to the user-provided CSV file to be validated.}

\item{sample_type}{Character string. The type of sample being validated. Expected values
are "micronix" or "cryovial".}

\item{user_action}{Character string. The action being performed by the user,
e.g. "upload".}

\item{database}{Character string. The database connection or specification for validation.
Default is the system environment variable "SDB_PATH".}
}
\value{
A data frame. The validated and potentially modified user data if validation
passes. If validation does not pass, an error will be raised.
}
\description{
The \code{validate_specimens} function serves as a main driver for validating specimens
data, both for Micronix and Cryovial specimens. The process involves several key steps:
}
\details{
\enumerate{
\item Formatting the provided user CSV based on the sample type and action with \code{format_user_csv}.
\item Validation based on the sample type using either \code{validate_micronix} or \code{validate_cryovial}.
\item If the user action is "upload", dates in the data will be further validated using \code{validate_dates_with_tokens}.
\item Any errors during the validation steps are accumulated and, if present, will trigger an
error message, stopping the process.
\item If unknown date tokens are found during validation, these are handled using the
\code{handle_unknown_date_tokens} function.
}

At the end of the validation process, the user data will be either ready for upload to the
database or returned to the user with errors.
}
\seealso{
\code{\link{format_user_csv}}, \code{\link{validate_micronix}},
\code{\link{validate_cryovial}}, \code{\link{validate_dates_with_tokens}},
\code{\link{handle_unknown_date_tokens}}
}
\keyword{validation}
