% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_strains}
\alias{validate_strains}
\title{Validate Strains Data}
\usage{
validate_strains(con, user_data)
}
\arguments{
\item{con}{The database connection.}

\item{user_data}{A data frame containing strain data to validate.}
}
\value{
An ErrorDataList object containing validation errors, if any.
}
\description{
This function conducts a series of validation checks on strain data.
The function will return any errors encountered during validation.
}
\keyword{validation}
