% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_study_reference}
\alias{validate_study_reference}
\title{Validate study references}
\usage{
validate_study_reference(con, table_name, row_number_col, study_short_code_col)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The name of the row number column.}

\item{study_short_code_col}{The name of the study short code column.}
}
\value{
ErrorData object indicating any studies not found.
}
\description{
This function checks if the study reference provided in the dataset exists in the database.
}
\keyword{validation}
