% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_study_reference_db}
\alias{validate_study_reference_db}
\title{Validate study references (Database Version)}
\usage{
validate_study_reference_db(
  con,
  table_name,
  row_number_col,
  study_short_code_col,
  controls = FALSE,
  error_if_exists = FALSE
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the formatted CSV table in the database.}

\item{row_number_col}{The name of the row number column.}

\item{study_short_code_col}{The name of the study short code column.}

\item{controls}{Whether are not we are validating controls. Controls have 'Batches' instead of studies.}

\item{error_if_exists}{Whether to indicate an error if a study / batch is already in the database.}
}
\value{
ErrorData object indicating any studies not found.
}
\description{
This function checks if the study reference provided in the dataset exists in the database.
Note: This function assumes that the data is already present in the database.
}
\keyword{validation}
