% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validate_study_short_code}
\alias{validate_study_short_code}
\title{Validate that study does not have a date for it's short code}
\usage{
validate_study_short_code(con, table_name, row_number, study_short_code_col)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the formatted CSV table in the database.}

\item{study_short_code_col}{The name of the study short code column.}

\item{row_number_col}{The name of the row number column.}
}
\description{
This function helps to validate that a study does not have a date for it's short code, which is reserved for control batches.
}
