% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_whole_blood_create}
\alias{validate_whole_blood_create}
\title{Validate Whole Blood Create}
\usage{
validate_whole_blood_create(whole_blood_test)
}
\arguments{
\item{whole_blood_test}{The utility function for performing validation checks.}
}
\value{
A list containing validation errors, if any.
}
\description{
Conducts specific validation checks for creating whole blood controls.
}
\keyword{validation}
