% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_whole_blood_extraction}
\alias{validate_whole_blood_extraction}
\title{Validate Whole Blood Extraction}
\usage{
validate_whole_blood_extraction(whole_blood_test)
}
\arguments{
\item{whole_blood_test}{The utility function for performing validation checks.}
}
\value{
A list containing validation errors, if any.
}
\description{
Conducts specific validation checks for extracting whole blood controls.
}
\keyword{validation}
