% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateControls.R
\name{validate_whole_blood_tube_exists}
\alias{validate_whole_blood_tube_exists}
\title{Validate Cryovial Tube Exists}
\usage{
validate_whole_blood_tube_exists(
  con,
  table_name,
  row_number_col,
  barcode_position_col,
  box_name_col,
  box_barcode_col,
  control_uid_col,
  batch_col
)
}
\value{
A list containing validation errors, if any.
}
\description{
This function checks that a cryovial exists by position, boxname, and barcode and box barcode, if provided.
}
\keyword{validation}
