% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{ValidationErrorCollection}
\alias{ValidationErrorCollection}
\title{ValidationErrorCollection class}
\description{
This class manages a collection of ErrorData objects.
It can be used to add new ErrorData objects, and to export the collection of errors to a CSV file.
}
\keyword{error-handling}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error_data_list}}{A list of ErrorData objects representing the collection of errors.}

\item{\code{user_data}}{The original data frame where the errors were found.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidationErrorCollection-new}{\code{ValidationErrorCollection$new()}}
\item \href{#method-ValidationErrorCollection-add_error}{\code{ValidationErrorCollection$add_error()}}
\item \href{#method-ValidationErrorCollection-merge_data_list}{\code{ValidationErrorCollection$merge_data_list()}}
\item \href{#method-ValidationErrorCollection-to_csv}{\code{ValidationErrorCollection$to_csv()}}
\item \href{#method-ValidationErrorCollection-length}{\code{ValidationErrorCollection$length()}}
\item \href{#method-ValidationErrorCollection-clone}{\code{ValidationErrorCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-new}{}}}
\subsection{Method \code{new()}}{
This method receives a list of ErrorData objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$new(errors = list(), user_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errors}}{The list of ErrorData objects to be added.}

\item{\code{user_data}}{The user data where the error data came from.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-add_error"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-add_error}{}}}
\subsection{Method \code{add_error()}}{
This method adds an ErrorData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$add_error(error_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error_data}}{The ErrorData object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-merge_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-merge_data_list}{}}}
\subsection{Method \code{merge_data_list()}}{
This method is used to merge the error data from another ErrorDataList into this one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$merge_data_list(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{The other ErrorDataList object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-to_csv}{}}}
\subsection{Method \code{to_csv()}}{
This method is used to export the errors to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$to_csv(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The name of the CSV file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-length"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-length}{}}}
\subsection{Method \code{length()}}{
This method returns the number of ErrorData objects in the list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
