% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{add_new_blood_spot_collections}
\alias{add_new_blood_spot_collections}
\title{Add New Blood Spot Collections}
\usage{
add_new_blood_spot_collections(
  df.payload,
  con,
  count_col,
  dbs_control_sheet_id
)
}
\arguments{
\item{df.payload}{A dataframe containing payload data.}

\item{con}{A database connection object.}

\item{count_col}{Column name in df.payload for count.}

\item{dbs_control_sheet_id}{Column name in df.payload for DBS control sheet ID.}
}
\value{
Result from the dbAppendTable indicating if the collections were added successfully.
}
\description{
Add New Blood Spot Collections
}
