% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAndDeleteSamples.R
\name{ArchiveAndDeleteSamples}
\alias{ArchiveAndDeleteSamples}
\title{Archive and Delete EPPIcenter Wetlab Samples}
\usage{
ArchiveAndDeleteSamples(operation, filters)
}
\description{
Archive and Delete EPPIcenter Wetlab Samples
}
