% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAndDeleteSamples.R
\name{ArchiveAndDeleteSamples}
\alias{ArchiveAndDeleteSamples}
\title{Archive and Delete Wetlab EPPIcenter Samples}
\usage{
ArchiveAndDeleteSamples(operation, data, comment, status, verification = TRUE)
}
\arguments{
\item{operation}{A string specifying the type of operation to perform. Valid operations are \code{archive} and \code{delete}.

\code{archive} should be used to delete all external data from the sample (cryovial_tube/paper/rdt information, container information, storage location)
but preserve internal data (sample id, specimen information, study subject information).
This operation can be used for example when a micronix cryovial_tube is emptied or rancid and needs to be discarded. \cr
\code{delete} should be used to remove all traces of the sample (internal and external data).
Deletions are recursive in the sense that they if sample ids associated with samples from an entire cryovial_box are deleted then the cryovial_box is automatically deleted.}

\item{verification}{A logical value. If set to TRUE user verification of the operation is required. If set to false user verification is bypassed. Default is set to TRUE.}

\item{sample_id}{A numeric vector of a one or more sample IDs (also referred to as storage container IDs). Can be accessed using sampleDB::SearchSamples().}
}
\description{
Archive and Delete Wetlab EPPIcenter Samples
}
\examples{
\dontrun{
ArchiveAndDeleteSamples(operation = "delete", sample_id = c(7:10))
ArchiveAndDeleteSamples(operation = "archive", sample_id = 100)
}
}
