% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{convert_density_representations}
\alias{convert_density_representations}
\title{Convert Density Representations to Real Numbers}
\usage{
convert_density_representations(user_data, density_col)
}
\arguments{
\item{user_data}{A dataframe containing user data with density values.}

\item{density_col}{A string specifying the column name of the density values in \code{user_data}.}
}
\value{
A dataframe with the converted density values.
}
\description{
This function processes a dataframe to convert density values that might be in various
representations (e.g., "10K", "10k") into their real number equivalents.
For instance, "10K" or "10k" will be converted to 10000.
}
\examples{
user_df <- data.frame(Density = c("1", "10k", "100K"))
convert_density_representations(user_df, "Density")

}
