% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSpecimens.R
\name{.UploadSpecimens}
\alias{.UploadSpecimens}
\title{Update the number of extracted DNA controls after extraction}
\usage{
.UploadSpecimens(upload_data, sample_type_id, conn)
}
\arguments{
\item{user_data}{A dataframe of SampleDB Upload data.}

\item{database}{Path to sampleDB database}
}
\description{
\code{UpdateActiveControlCount()}
}
