% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{format_labels}
\alias{format_labels}
\title{Format labels for compositions}
\usage{
format_labels(compositions)
}
\arguments{
\item{compositions}{A dataframe containing the compositions to be labeled.
The dataframe should contain the columns \code{legacy}, \code{label}, and \code{index}.}
}
\value{
A character vector containing the formatted labels.
}
\description{
This function formats the labels for the given compositions based on the legacy status and index.
}
\examples{
df <- data.frame(legacy = c(1, 0), label = c("S1", "S2"), index = c(1, 2))
format_labels(df)
}
