% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_control_file_columns}
\alias{get_control_file_columns}
\title{Get user action fields for controls}
\usage{
get_control_file_columns(control_type, user_action)
}
\arguments{
\item{control_type}{The type of control}

\item{user_action}{The user action to get fields for}
}
\value{
A FileColumnAttributes object with fields for the user action for controls
}
\description{
This function gets all the fields for a given user action related to controls.
}
\keyword{appdata}
