% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{get_identifiers_from_database}
\alias{get_identifiers_from_database}
\title{Extract identifiers from the database}
\usage{
get_identifiers_from_database(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing unique identifiers for each composition in the database.
}
\description{
This function queries a database to collect and process data, generating a unique identifier based on strains and percentages.
}
