% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_reference_file_columns}
\alias{get_reference_file_columns}
\title{Retrieve File Column Attributes for a Given Category and Reference Type}
\usage{
get_reference_file_columns(category, reference_type = NULL)
}
\arguments{
\item{category}{A string specifying the category of reference, e.g., "locations", "strains".}

\item{reference_type}{An optional string specifying the specific type within the category, e.g., "minus20". Default is NULL.}
}
\value{
A \code{FileColumnAttributes} object populated with the fields related to the given category and reference type.
}
\description{
This function gets all the fields for a given user action related to references
and returns a \code{FileColumnAttributes} object populated with those fields.
It supports different reference categories in the JSON, such as locations, strains, compositions, etc.
Depending on the category and the reference type (if provided),
the function will fetch fields categorized as \code{required}, \code{optional}, or directly if no sub-categories are present.
}
\examples{
\dontrun{
# For categories with a specific type
attributes <- get_reference_file_columns("locations", "minus20")

# For categories without a specific type
attributes <- get_reference_file_columns("strains")
}

}
