% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppData.R
\name{get_sample_file_columns}
\alias{get_sample_file_columns}
\title{Retrieve ColumnData for a given sample type and action.}
\usage{
get_sample_file_columns(
  sample_type,
  action,
  file_type = "na",
  container_type = NULL,
  config_yml = Sys.getenv("SDB_CONFIG"),
  samples_file = "samples.json"
)
}
\arguments{
\item{sample_type}{Type of the sample for which the columns are needed.}

\item{action}{Desired action.}

\item{file_type}{The file type being used. Default is 'na'.}

\item{config_yml}{(default: Sys.getenv("SDB_PATH")) Path to the application config file}

\item{samples_file}{(default: "samples.json") Path to the JSON file containing samples data.}
}
\value{
A ColumnData S3 object.
}
\description{
This function provides the ColumnData S3 object for a specific sample type and action.
}
