% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveSamples.R
\name{MoveSamples}
\alias{MoveSamples}
\title{Move Wetlab Samples}
\usage{
MoveSamples(sample_type, move_data)
}
\arguments{
\item{sample_type}{A string specifying the type of samples that are being moved. Options include: \code{micronix}, \code{cryovial}, \code{rdt} and \code{paper}}

\item{move_data}{A list of SampleDB move dataframes, where the name of each dataframe item is the container that the samples are in after the move.}
}
\description{
\code{MoveSamples()} can be used to move existing wetlab samples from:
\enumerate{
\item Between two (or more) existing containers
\item From one (or more) existing container into one (or more) existing container
}
}
\examples{
\dontrun{
move_data <- list("move_csv1_container_name" = dataframe(), "move_csv2_container_name" = dataframe())
MoveSamples(sample_type = "micronix", move_data = move_data)
}
}
