% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveSamples.R
\name{MoveSamples}
\alias{MoveSamples}
\title{Move Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
MoveSamples(sample_type, move_files)
}
\arguments{
\item{sample_type}{A string specifying the type of samples that are being moved. Options include: \code{micronix}, \code{cryovial}, \code{rdt} and \code{paper}}

\item{move_files}{A list of paths to SampleDB Move CSV files. The items in the \code{move_files} list must be named after the container associate with each Move CSV file.

The basic structure of a Move CSV file is shown below.\tabular{lll}{
   row \tab column \tab label \cr
   A \tab 0 \tab xxx1 \cr
   A \tab 1 \tab xxx2 \cr
}


Accepted names for the \code{row} column include: \code{row}, \code{Row}, and \code{LocationRow}. \cr
Accepted names for the \code{column} column include: \code{column}, \code{Column}, and \code{LocationColumn}. \cr
Also accepted is a single column named \code{Position} that represents the both the \code{row} column and the \code{column} column.
If the \code{Position} column is used then the \code{row} and \code{column} columns can be ommited.}
}
\description{
Move Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
\dontrun{
move_files <- list("move_csv1_container_name" = "~/path/to/move_csv1.csv", "move_csv2_container_name" = "~/path/to/move_csv2.csv")
MoveSamples(sample_type = "micronix", move_files = move_files)
}
}
