% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadControls.R
\name{process_blood_spot_collection_data}
\alias{process_blood_spot_collection_data}
\title{Process Blood Spot Collection Data}
\usage{
process_blood_spot_collection_data(df.payload, con, count_col)
}
\arguments{
\item{df.payload}{A dataframe containing payload data.}

\item{con}{A database connection object.}

\item{count_col}{Count column.}
}
\value{
A dataframe with processed blood spot collection details.
}
\description{
Process Blood Spot Collection Data
}
