% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{SearchSamples}
\alias{SearchSamples}
\title{Search for Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
SearchSamples(
  sample_type = NULL,
  sample_label = NULL,
  container_name = NULL,
  study_subject = NULL,
  specimen_type = NULL,
  study = NULL,
  collection_dates = NULL,
  archived = NULL,
  freezer = NULL,
  study_subject.file = FALSE,
  return_sample_ids = FALSE
)
}
\arguments{
\item{sample_type}{c("micronix", "cryovile", "rdt", "paper")}

\item{container_name}{list(micronix.container_name = "", cryovile.container_name = "", rdt.container_name = "", paper.container_name = "")}

\item{specimen_type}{c("PLASMA", "DNA (BS)")}

\item{study}{c("PRISM", "STUDY")}

\item{collection_dates}{list(date.to = "", date.from = "")}

\item{archived}{TRUE/FALSE}

\item{study_subject.file}{TRUE or FALSE}

\item{return_sample_ids}{TRUE or FALSE}

\item{sample_labels}{list(micronix.label_name = "", cryovile.label_name = "", rdt.label_name = "", paper.label_name = "")}

\item{study_subjects}{c("subj_1", subj_2")}

\item{freezers}{list(location_name = "", level_I = "", level_II)}
}
\description{
Search for Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
SearchSamples(study = "MF", study_subject = "MF-20079")

NOTE: Unfortunately the whole database has to be read into memory inorder to search for something
Im sure there is a faster sql way to do this
}
