% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{SearchSamples}
\alias{SearchSamples}
\title{Search for Wetlab Samples in the EPPIcenter SampleDB database}
\usage{
SearchSamples(
  sample_type = NULL,
  sample_label = NULL,
  container_name = NULL,
  study_subject = NULL,
  specimen_type = NULL,
  study = NULL,
  collection_dates = NULL,
  status = NULL,
  state = NULL,
  freezer = NULL,
  return_sample_ids = FALSE
)
}
\arguments{
\item{sample_type}{A string specifying the type of EPPIcenter sample. (\code{micronix}, \code{cryovile}, \code{rdt} and/or \code{paper})}

\item{sample_label}{A list of vectors specifying the vector micronix, cryovile, rdt, and paper label names (\code{micronix.label_name}, \code{cryovile.label_name}, \code{rdt.label_name} and/or \code{paper.label_name})}

\item{container_name}{A list of vectors specifying the vector micronix, cryovile, rdt, and paper container names (\code{micronix.container_name}, \code{cryovile.container_name}, \code{rdt.container_name} and/or \code{paper.container_name})}

\item{study_subject}{A study subjects string or a vector of study subject strings. If \code{study_subject.file} is TRUE the path to a .csv file containing one column named study_subject can be uploaded and used to search the database for study subjects.}

\item{specimen_type}{A specimen type string or a vector of specimen type strings.}

\item{study}{A study short code string or a vector of study short code strings.}

\item{collection_dates}{A list of date values strings (\code{date.to} and \code{date.from}) that can be used to filter EPPIcenter samples}

\item{freezer}{A list specifying the vector \code{location_name}, \code{level_I}, and/or\code{level_II}}

\item{return_sample_ids}{A logical value. Setting \code{return_sample_ids} to \code{TRUE} means \code{SearchSamples} returns sample ids as well as search results. Setting \code{return_sample_ids} to \code{FALSE} means \code{SearchSamples} returns only search results. Default value is \code{FALSE}.}

\item{archived}{A logical value. \code{TRUE} filters for archived samples and \code{FALSE} filters for active samples}
}
\description{
Search for Wetlab Samples in the EPPIcenter SampleDB database
}
\examples{
\dontrun{
SearchSamples(study = "KAM06", study_subject = "subject_1")
}
}
