% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchSamples.R
\name{SearchSamples}
\alias{SearchSamples}
\title{Search for Wetlab Samples in the EPPIcenter sampleDB database}
\usage{
SearchSamples(filters, study_subject.file = F)
}
\arguments{
\item{filters}{A list containing micronix barcodes, plate name, study code, location and/or specimen_type}

\item{study_subject.file}{TRUE or FALSE}
}
\description{
Search for Wetlab Samples in the EPPIcenter sampleDB database
}
\examples{
SearchSamples(filters = list(name.plate = c("100","101"), name.location = c("Left -20 Freezer")))
SearchSamples(filters = list(file.barcodes = "/path/to/barcodes.csv", name.plate = "dummy_name", name.study = "dummy_study", name.location = "fridge", name.specimen_type = "specimen_type1"))
}
