% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{set_user_file_header}
\alias{set_user_file_header}
\title{Set the header row for the user CSV file}
\usage{
set_user_file_header(user_file, required_columns)
}
\arguments{
\item{user_file}{The user's CSV file as a data frame.}

\item{required_columns}{A character vector of required column names.}
}
\value{
A data frame with the appropriate header row set.
}
\description{
This function identifies the valid header row in the user file and sets it as the header.
If no valid header row is found, a formatting error is thrown.
}
