% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateReferences.R
\name{UpdateReferences}
\alias{UpdateReferences}
\title{Update References in the EPPIcenter sampleDB database}
\usage{
UpdateReferences(reference, operation, information)
}
\arguments{
\item{reference}{One of the following strings: freezer, specimen_type, study}

\item{operation}{One of the following strings: add, modify, delete}

\item{information}{A key value pair that matches the specified reference and operation}
}
\description{
Update References in the EPPIcenter sampleDB database
}
\examples{

UpdateReferences(reference = "study", operation = "modify", information = list(OldStudyShortCode = "C", NewStudyTitle = "AA", NewLeadPerson = "EE"))
UpdateReferences(reference = "study", operation = "delete", information = list(DeleteStudyShortCode = "MNP"))
}
