% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSamples.R
\name{UploadSamples}
\alias{UploadSamples}
\title{Upload EPPIcenter Wetlab Samples to the sampleDB database}
\usage{
UploadSamples(csv.upload, name.plate, location)
}
\arguments{
\item{csv.upload}{An UploadCSV}

\item{name.plate}{A plate name}

\item{location}{A freezer in the lab}
}
\description{
Upload EPPIcenter Wetlab Samples to the sampleDB database
}
\examples{
UploadSamples(csv.upload = "/path/to/UploadCSV.csv", name.plate = "dummy_name", location = "Left -20 Freezer")
}
