% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSamples.R
\name{UploadSamples}
\alias{UploadSamples}
\title{Upload EPPIcenter Wetlab Samples to the sampleDB database}
\usage{
UploadSamples(type, csv.upload, container, list.location)
}
\arguments{
\item{csv.upload}{An UploadCSV}

\item{container}{A plate name}

\item{list.location}{A freezer in the lab + a level I in that freezer + a level II in that freezer}
}
\description{
Upload EPPIcenter Wetlab Samples to the sampleDB database
}
\examples{
UploadSamples(csv.upload = "/path/to/UploadCSV.csv", container = "dummy_name", location = list.location = list(location_name = "Freezer A", level_I = "dummy.levelI", level_II = "dummy.levelII"))
}
