% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSamples.R
\name{UploadSamples}
\alias{UploadSamples}
\title{Upload EPPIcenter Wetlab Samples to the SampleDB database}
\usage{
UploadSamples(sample_type, upload_file, container_name, freezer)
}
\arguments{
\item{sample_type}{A string specifying the type of samples that are being uploaded Options include: \code{micronix}, \code{cryovial}, \code{rdt} and \code{paper}}

\item{upload_file}{A path string for the SampleDB Upload CSV file.

\strong{Upload Micronix CSV structure (vision mate + no collection date)}\tabular{llllll}{
   LocationRow \tab LocationColumn \tab TubeCode \tab study_subject_id \tab specimen_type \tab study_short_code \cr
   A \tab 0 \tab xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \cr
   A \tab 1 \tab xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \cr
}


\strong{Upload Micronix CSV structure (vision mate + collection date)}\tabular{lllllll}{
   LocationRow \tab LocationColumn \tab TubeCode \tab study_subject_id \tab specimen_type \tab study_short_code \tab collection_date \cr
   A \tab 0 \tab xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
   A \tab 1 \tab xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
}


\strong{Upload Micronix CSV structure (traxer + no collection date)}\tabular{lllllllll}{
   Position \tab Tube ID \tab Status \tab Error Count \tab Rack ID \tab Date \tab study_subject_id \tab specimen_type \tab study_short_code \cr
   A0 \tab xxx1 \tab  \tab  \tab  \tab  \tab subject_1 \tab PLASMA \tab KAM06 \cr
   A1 \tab xxx2 \tab  \tab  \tab  \tab  \tab subject_2 \tab PLASMA \tab KAM06 \cr
}


\strong{Upload Micronix CSV structure (traxer + collection date)}\tabular{llllllllll}{
   Position \tab Tube ID \tab Status \tab Error Count \tab Rack ID \tab Date \tab study_subject_id \tab specimen_type \tab study_short_code \tab collection_date \cr
   A0 \tab xxx1 \tab  \tab  \tab  \tab  \tab subject_1 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
   A1 \tab xxx2 \tab  \tab  \tab  \tab  \tab subject_2 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
}


\strong{Upload Cryovial CSV structure (no collection date)}\tabular{llllll}{
   row \tab column \tab label \tab study_subject_id \tab specimen_type \tab study_short_code \cr
   A \tab 0 \tab xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \cr
   A \tab 1 \tab xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \cr
}


\strong{Upload Cryovial CSV structure (collection date)}\tabular{lllllll}{
   row \tab column \tab label \tab study_subject_id \tab specimen_type \tab study_short_code \tab collection_date \cr
   A \tab 0 \tab xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
   A \tab 1 \tab xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
}


\strong{Upload RDT CSV structure (no collection date)}\tabular{llll}{
   label \tab study_subject_id \tab specimen_type \tab study_short_code \cr
   xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \cr
   xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \cr
}


\strong{Upload RDT CSV structure (collection date)}\tabular{lllll}{
   label \tab study_subject_id \tab specimen_type \tab study_short_code \tab collection_date \cr
   xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
   xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
}


\strong{Upload Paper CSV structure}\tabular{llll}{
   label \tab study_subject_id \tab specimen_type \tab study_short_code \cr
   xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \cr
   xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \cr
}


\strong{Upload Paper CSV structure}\tabular{lllll}{
   label \tab study_subject_id \tab specimen_type \tab study_short_code \tab collection_date \cr
   xxx1 \tab subject_1 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
   xxx2 \tab subject_2 \tab PLASMA \tab KAM06 \tab 2022-02-11 \cr
}}

\item{container_name}{A string specifying the name of the container the samples are in. Names must be unique within each sample type.}

\item{freezer}{A list specifying the freezer used to store samples. \cr
Required items in the freezer list are \code{location_name}, \code{level_I} and \code{level_II}.
If the freezer type is \verb{minus eighty} then \code{level_I} and \code{level_II} items specify the rack and position, respecively.
If the freezer type is \verb{minus twenty} then \code{level_I} and \code{level_II} items specify the shelf and basket, respecively.}
}
\description{
\code{UploadSamples()} can be used to upload wetlab samples to the sampleDB database.
Currently the type of wetlab samples supported is only \code{micronix}. \code{cryovial}, \code{rdt} and \code{paper} sample uploads will appear in the next version.
}
\examples{
\dontrun{
UploadSamples(upload_file = "/path/to/UploadCSV.csv", container_name = "dummy_name", location = freezer = list(location_name = "Freezer A", level_I = "dummy.levelI", level_II = "dummy.levelII"))
}
}
