% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{validate_and_format_control_file}
\alias{validate_and_format_control_file}
\title{Check if the user CSV file for control meets the requirements}
\usage{
validate_and_format_control_file(user_file, file_column_attr, bind_data = NULL)
}
\arguments{
\item{user_file}{The user's CSV file as a data frame.}

\item{file_column_attr}{A list containing attributes of file columns such as required, conditional, and optional columns.}

\item{bind_data}{Data that should be added to the data frame.}
}
\value{
A cleaned and checked user_file ready for further processing.
}
\description{
This function checks if the user CSV file for control meets specific upload requirements,
including necessary columns and formatting checks. Errors will be thrown if requirements are not met.
}
