% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCSV.R
\name{validate_and_format_specimen_file}
\alias{validate_and_format_specimen_file}
\title{Check if the user CSV file meets the requirements}
\usage{
validate_and_format_specimen_file(
  user_file,
  sample_type,
  user_action,
  file_column_attr,
  container_name = NULL,
  freezer_address = NULL
)
}
\arguments{
\item{user_file}{The user's CSV file as a data frame.}

\item{sample_type}{The sample type associated with the user's CSV file.}

\item{file_column_attr}{The file column attributes.}

\item{container_name}{The name of the container.}

\item{freezer_address}{The address of the freezer.}
}
\value{
A cleaned and checked user_file ready for further processing.
}
\description{
This function checks if the user CSV file meets specific upload requirements,
including necessary columns and formatting checks. Errors will be thrown if requirements are not met.
}
