% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSpecimens.R
\name{validate_matrix_container}
\alias{validate_matrix_container}
\title{Validate Matrix Container position}
\usage{
validate_matrix_container(
  con,
  user_data,
  row_number_col,
  container_name_col,
  container_position_col,
  container_tablename,
  error_if_exists
)
}
\arguments{
\item{con}{A database connection object.}

\item{user_data}{The name of the table where the user data is temporarily stored in the database.}

\item{row_number_col}{The column with the row number in the \code{user_data}.}

\item{error_if_exists}{Logical. If TRUE, an error is returned if the barcode exists in the database.}

\item{cryovial_col}{The column with Cryovial barcodes in the \code{user_data}.}

\item{cryovial_box_col}{The column with Cryovial box IDs in the \code{user_data}.}

\item{cryovial_box_barcode_col}{The column with Cryovial box barcodes in the \code{user_data}.}
}
\value{
An ErrorData object if Micronix position rules are violated, or NULL otherwise.
}
\description{
Validate Matrix Container position
}
