% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_references}
\alias{validate_references}
\title{Validate References}
\usage{
validate_references(database, user_data, reference_type, user_action)
}
\arguments{
\item{database}{A database connection or reference.}

\item{reference_type}{The type of reference ('strains' or 'compositions').}

\item{user_action}{The user action taken on the control specimen ('extraction' or 'create').}

\item{formatted_csv}{The CSV data formatted for validation.}
}
\value{
nothing
}
\description{
Validate References
}
