% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReferences.R
\name{validate_strains_in_database}
\alias{validate_strains_in_database}
\title{Validate strains recorded in the database}
\usage{
validate_strains_in_database(
  con,
  table_name,
  row_number_col,
  strain_col,
  error_select,
  error_if_exists = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{table_name}{The name of the table in the database that contains the user data.}

\item{row_number_col}{The column name for row numbers.}

\item{strain_col}{The column name for strains.}

\item{error_select}{Select which column to report with errors. The column used here will depend on what data you want the user to see regarding strains.}

\item{error_if_exists}{Logical. If TRUE, checks if the strain exists in the database. If FALSE, checks if it's absent.}
}
\value{
An instance of the ErrorData class or NULL.
}
\description{
Validate strains recorded in the database
}
\keyword{validation}
