% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{ValidationErrorCollection}
\alias{ValidationErrorCollection}
\title{ValidationErrorCollection class}
\description{
This class manages a collection of ErrorData objects.
It can be used to add new ErrorData objects, and to export the collection of errors to a CSV file.
}
\keyword{error-handling}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error_data_list}}{A list of ErrorData objects representing the collection of errors.}

\item{\code{user_data}}{The original data frame where the errors were found.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidationErrorCollection-new}{\code{ValidationErrorCollection$new()}}
\item \href{#method-ValidationErrorCollection-add_error}{\code{ValidationErrorCollection$add_error()}}
\item \href{#method-ValidationErrorCollection-get_error_details_by_index}{\code{ValidationErrorCollection$get_error_details_by_index()}}
\item \href{#method-ValidationErrorCollection-length}{\code{ValidationErrorCollection$length()}}
\item \href{#method-ValidationErrorCollection-get_combined_error_data}{\code{ValidationErrorCollection$get_combined_error_data()}}
\item \href{#method-ValidationErrorCollection-clone}{\code{ValidationErrorCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-new}{}}}
\subsection{Method \code{new()}}{
This method receives a list of ErrorData objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$new(errors = list(), user_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errors}}{The list of ErrorData objects to be added.}

\item{\code{user_data}}{The user data where the error data came from.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-add_error"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-add_error}{}}}
\subsection{Method \code{add_error()}}{
This method adds an ErrorData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$add_error(error_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error_data}}{The ErrorData object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-get_error_details_by_index"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-get_error_details_by_index}{}}}
\subsection{Method \code{get_error_details_by_index()}}{
Gets an error by index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$get_error_details_by_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The index to specify which error.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-length"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-length}{}}}
\subsection{Method \code{length()}}{
This method returns the number of ErrorData objects in the list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-get_combined_error_data"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-get_combined_error_data}{}}}
\subsection{Method \code{get_combined_error_data()}}{
This method combines the error details with the user data and returns it as a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$get_combined_error_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
